Split the render path and pointer input out of app/mod.rs
3422 lines holding the state machine, both halves of the render path and
every hit test. Now 1062, with the rest in modules named for what they do:

- `app::render::prepare` — rasterization, one `prepare_*` per family
- `app::render::compose` — the draw list, pure texture-copy bookkeeping
- `app::render::geometry` — what both halves measure against (scroll extent,
  tile cropping, modal card rects), so they cannot derive it differently
- `app::pointer` — Magic Remote hover and click, against the same
  `app::view` rects the painter draws with

Pure moves: only two signatures rustfmt rewrapped, plus the visibility
widening the split forces. The `use crate::app::*` in each is deliberate —
these are `impl App` blocks lifted out of that module and they read the same
private tuning constants.

Deviates from B1 in docs/UI-Library-Hardening.md, which had these become
free functions over a completed `RenderInput`. `RenderInput` is five fields
after several families; finishing it means a ~40-field mirror of `App`, and
its stated payoff was a PNG-diff test harness this project does not use
(it verifies on device). The file split is the part that was actually worth
having. `RenderInput` stays for the families already on it.
