+++ src/app/mod.rs
@@ -7,6 +7,7 @@
 pub(crate) mod hero;
 pub(crate) mod hosts;
 pub(crate) mod menu;
+pub(crate) mod pointer;
 pub(crate) mod render;
 pub(crate) mod render_input;
 pub(crate) mod state;
@@ -19,7 +20,7 @@ use anyhow::Result;
 use tiny_skia::Pixmap;
 
 use crate::app::hosts::HostEntry;
-use crate::app::render::key::{ModalFocusKey, ModalShellKey, ScrollContentKey};
+use crate::app::render::key::ModalShellKey;
 use crate::app::render::tile;
 use crate::app::state::addhost::AddHostState;
 use crate::core::event::MenuEvent;
@@ -28,8 +29,7 @@ use crate::core::model::GameEntry;
 pub use crate::core::screen::{HomeFocus, PairingFocus, Screen};
 use crate::services::store::{self, KnownHost, Settings};
 use crate::ui;
-use crate::ui::cache::{self, TileStore};
-use crate::ui::render::{DrawCmd, TileId};
+use crate::ui::render::TileId;
 use crate::ui::Painter;
 
 /// Rows beyond viewport kept rasterized (prevents scroll stalls).
@@ -970,667 +970,6 @@ impl App {
         }
         animating
     }
-    // ---------------------------------------------------------------- mouse --
-
-    /// Title and subtitle of the address form, which serves both Add host and Edit
-    /// address — the only difference between the two screens.
-    fn address_copy(&self) -> (&'static str, String) {
-        match self.screen {
-            Screen::EditHost => {
-                let name = self
-                    .edit_host_index
-                    .and_then(|i| self.entries.get(i))
-                    .map_or_else(String::new, |e| e.name().to_string());
-                (view::addhost::EDIT_TITLE, view::addhost::edit_subtitle(&name))
-            }
-            _ => (view::addhost::ADD_TITLE, view::addhost::ADD_SUBTITLE.to_string()),
-        }
-    }
-
-    /// Handed to `SDL_SetTextInputRect` by the render loop.
-    pub fn address_field_rect(&self, screen_w: u32, screen_h: u32, fonts: &ui::text::Fonts) -> Rect {
-        let (_, subtitle) = self.address_copy();
-        view::addhost::field_rect(screen_w, screen_h, fonts, &subtitle, self.keyboard_shown)
-    }
-
-    /// Updates focus/hover to whatever the Magic Remote's pointer is over, returning
-    /// whether that changed anything visible — Magic Remote pointer mode fires
-    /// `MouseMotion` continuously while moving, so callers redraw only when this is
-    /// `true` rather than on every event.
-    pub fn handle_mouse_motion(
-        &mut self,
-        x: i32,
-        y: i32,
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::text::Fonts,
-    ) -> bool {
-        let focus_changed = self.hover_focus_at(x, y, screen_w, screen_h, fonts);
-        // Parity with the D-pad: a hover that moves modal focus replays the focus-pop zoom
-        // (and shows the new row's caption). Home drives its own `focus_anim` instead, so
-        // it's excluded. An open dropdown is excluded too — hover there only moves the
-        // option cursor, so popping the parent row (as the D-pad also declines to) is wrong.
-        if focus_changed && self.dropdown.is_none() && !matches!(self.screen, Screen::Home) {
-            self.modal_focus_anim = Some(Instant::now());
-        }
-        let close_changed = self.hover_close_at(x, y, screen_w, screen_h, fonts);
-        focus_changed || close_changed
-    }
-
-    /// Button index under `(x, y)` for a two-button confirm modal with `subtitle`, or
-    /// `None` off both buttons — every confirm modal's hover arm shares this, against the
-    /// same `confirm_dialog_layout` geometry the modal is drawn with.
-    fn confirm_button_at(
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::text::Fonts,
-        subtitle: &str,
-        x: i32,
-        y: i32,
-    ) -> Option<usize> {
-        let (_, content) = ui::tiles::confirm_dialog_layout(screen_w, screen_h, fonts, subtitle);
-        ui::tiles::confirm_button_at(content, x, y)
-    }
-
-    /// Rect of confirm button `index` for a two-button modal with `subtitle` — the shared
-    /// geometry the focused-button tile and its hit-rect are positioned against.
-    fn confirm_focus_button_rect(
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::text::Fonts,
-        subtitle: &str,
-        index: usize,
-    ) -> Rect {
-        let (_, content) = ui::tiles::confirm_dialog_layout(screen_w, screen_h, fonts, subtitle);
-        ui::widgets::confirm_button_rect(content, index)
-    }
-
-    /// Moves the positional focus/selection onto whatever interactive element sits
-    /// under the pointer, so the Magic Remote's pointer highlights elements on hover
-    /// exactly where a click would land. Returns whether the selection actually
-    /// moved. Hovering empty space (gaps, row padding, the area between rows) leaves
-    /// the current selection put rather than clearing it, so a resting pointer never
-    /// fights the D-pad.
-    fn hover_focus_at(&mut self, x: i32, y: i32, screen_w: u32, screen_h: u32, fonts: &ui::text::Fonts) -> bool {
-        // An open dropdown overlays the row list — hover moves its option cursor and
-        // nothing behind it. Shared by whichever screen owns the dropdown (Settings or
-        // Diagnostics), and uses the same overlay geometry the renderer draws against.
-        if let Some(i) = self.dropdown_option_at(x, y, screen_w, screen_h, fonts) {
-            let dd = self
-                .dropdown
-                .as_mut()
-                .expect("dropdown_option_at yields Some only when one is open");
-            let changed = dd.focused != i;
-            dd.focused = i;
-            return changed;
-        }
-        // A dropdown open but not hovered still swallows hover — the row list behind
-        // it must not take the selection.
-        if self.dropdown.is_some() {
-            return false;
-        }
-        match self.screen {
-            Screen::Home => {
-                // The ⋯ button sits inside its row, so it's tested first — same order
-                // as `handle_mouse_click`, so hover previews exactly what a click hits.
-                if let Some(idx) =
-                    view::sidebar::hit_test_menu_button(x, y, self.entries.len(), self.sidebar_len(), screen_h)
-                {
-                    return self.set_home_focus(HomeFocus::SidebarMenu(idx));
-                }
-                if let Some(idx) = view::sidebar::hit_test_row(x, y, self.sidebar_len(), screen_h) {
-                    return self.set_home_focus(HomeFocus::Sidebar(idx));
-                }
-                let available_w = screen_w.saturating_sub(ui::widgets::SIDEBAR_W);
-                let columns = view::home::grid_columns(available_w);
-                if let Some(idx) = view::home::hit_test_grid_card(
-                    x,
-                    y,
-                    columns,
-                    self.grid_len(columns),
-                    ui::widgets::SIDEBAR_W as i32,
-                    available_w,
-                    self.grid_scroll,
-                ) {
-                    // Padding after a partial pinned row isn't a real card — nothing to land on.
-                    if self.is_grid_card(idx, columns) {
-                        return self.set_home_focus(HomeFocus::Grid(idx));
-                    }
-                }
-                false
-            }
-            // Dropdown case already handled above.
-            Screen::Settings => {
-                let Some(row) = self.settings_row_at(x, y, screen_w, screen_h) else {
-                    return false;
-                };
-                let changed = self.settings_focused != row;
-                self.settings_focused = row;
-                changed
-            }
-            Screen::HostMenu => {
-                let rows = self.host_menu_actions().len();
-                let Some((_, content)) = self.modal_list_geometry(screen_w, screen_h, fonts) else {
-                    return false;
-                };
-                let Some(i) = (0..rows).find(|&i| ui::widgets::focus_row_rect(content, i).contains_point((x, y)))
-                else {
-                    return false;
-                };
-                let row = ui::widgets::focus_row_rect(content, i);
-                let dots =
-                    self.host_menu_row_has_dots() && ui::widgets::sidebar_menu_button_rect(row).contains_point((x, y));
-                let changed = self.menu_focused != i || self.host_menu_dots != dots;
-                self.menu_focused = i;
-                self.host_menu_dots = dots;
-                changed
-            }
-            // Identical row-list geometry; only which focus field they carry differs.
-            Screen::Diagnostics | Screen::Experimental | Screen::CursorSettings => {
-                let Some(row) = self.modal_list_row_at(x, y, screen_w, screen_h, fonts) else {
-                    return false;
-                };
-                let focused = match self.screen {
-                    Screen::Diagnostics => &mut self.diagnostics_focused,
-                    Screen::CursorSettings => &mut self.cursor_settings_focused,
-                    _ => &mut self.experimental_focused,
-                };
-                let changed = *focused != row;
-                *focused = row;
-                changed
-            }
-            Screen::Pairing => {
-                let card = view::pairing::card_rect(screen_w, screen_h, fonts);
-                if view::pairing::request_button_rect(card, fonts).contains_point((x, y)) {
-                    let changed = self.pairing_focus != PairingFocus::RequestAccess;
-                    self.pairing_focus = PairingFocus::RequestAccess;
-                    changed
-                } else {
-                    false
-                }
-            }
-            // Two-button confirm modals (Forget/SendLogs/Wake — the same modal type as the
-            // in-stream Disconnect dialog): hovering a button focuses it, so the pointer can
-            // pick action-vs-Cancel, not just confirm whatever the D-pad last focused. All
-            // three share `confirm_button_at`; only the focus field they set differs.
-            Screen::ForgetHost => {
-                let name = self
-                    .host_menu_index
-                    .and_then(|i| self.entries.get(i))
-                    .map(HostEntry::name)
-                    .unwrap_or_default();
-                match Self::confirm_button_at(screen_w, screen_h, fonts, &view::forget::subtitle(name), x, y) {
-                    Some(i) => {
-                        let changed = self.host_menu_focused != i;
-                        self.host_menu_focused = i;
-                        changed
-                    }
-                    None => false,
-                }
-            }
-            Screen::SendLogs => {
-                match Self::confirm_button_at(screen_w, screen_h, fonts, view::sendlogs::SUBTITLE, x, y) {
-                    Some(i) => {
-                        let changed = self.send_logs_focused != i;
-                        self.send_logs_focused = i;
-                        changed
-                    }
-                    None => false,
-                }
-            }
-            // Only with a MAC — the no-MAC wake variant is a button-less message.
-            Screen::Wake => {
-                let Some(wake) = self.wake.as_ref().filter(|w| !w.mac.is_empty()) else {
-                    return false;
-                };
-                let Some(i) = Self::confirm_button_at(screen_w, screen_h, fonts, &view::wake::status_text(wake), x, y)
-                else {
-                    return false;
-                };
-                let wake = self.wake.as_mut().expect("filtered to Some above");
-                let changed = wake.focused != i;
-                wake.focused = i;
-                changed
-            }
-            // The finished/failed speed test shows the same two-button confirm row as
-            // ForgetHost et al., so hover picks apply-vs-Close there too. While the test
-            // is still running there are no buttons — `speed_test_buttons_rect` is only
-            // meaningful once `render_speed_test` draws them (Done/Failed).
-            Screen::SpeedTest
-                if matches!(
-                    self.speed_test,
-                    Some(crate::app::state::speedtest::SpeedTestState::Done { .. })
-                        | Some(crate::app::state::speedtest::SpeedTestState::Failed(_))
-                ) =>
-            {
-                let card = view::speedtest::card_rect(
-                    screen_w,
-                    screen_h,
-                    fonts,
-                    self.speed_test.as_ref(),
-                    &self.speed_test_name,
-                );
-                let content =
-                    view::speedtest::buttons_rect(card, fonts, self.speed_test.as_ref(), &self.speed_test_name);
-                match ui::tiles::confirm_button_at(content, x, y) {
-                    Some(i) => {
-                        let changed = self.speed_test_focused != i;
-                        self.speed_test_focused = i;
-                        changed
-                    }
-                    None => false,
-                }
-            }
-            // No positional focus to move: single-card info/entry modals (AddHost,
-            // EditHost, About, WakeSettings, PinLimit, running SpeedTest) and Settings
-            // with a dropdown open.
-            _ => false,
-        }
-    }
-
-    /// Sets `home_focus`, reporting whether it actually moved — the hover/click
-    /// helpers redraw only on a real change.
-    fn set_home_focus(&mut self, focus: HomeFocus) -> bool {
-        let changed = self.home_focus != focus;
-        self.home_focus = focus;
-        changed
-    }
-
-    /// The `(content viewport, pixel scroll offset)` an open dropdown anchors its
-    /// option overlay to, matching what `draw_list` renders so hit-testing lands
-    /// exactly where options are drawn. `None` for a screen with no dropdown.
-    /// `screen` is a param (not `self.screen`) so `draw_list`'s close-fade can pass
-    /// the screen it captured at `back()` time.
-    pub(crate) fn dropdown_geom(
-        &self,
-        screen: Screen,
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::text::Fonts,
-    ) -> Option<(Rect, i32)> {
-        match screen {
-            Screen::Settings => {
-                let (_, content) = view::settings::layout(&self.settings, screen_w, screen_h);
-                let stride = ui::widgets::focus_row_stride() as i32;
-                let total = menu::settings_row_count(&self.settings);
-                // Anchor to the animated offset so an open dropdown stays attached to
-                // its row while the list is still settling.
-                let px = self
-                    .modal_scroll_px
-                    .clamp(0, Self::max_scroll_px(total, stride, content.height()));
-                Some((content, px))
-            }
-            // Diagnostics doesn't scroll, so 0.
-            Screen::Diagnostics => Some((self.modal_list_geometry(screen_w, screen_h, fonts)?.1, 0)),
-            _ => None,
-        }
-    }
-
-    /// The Settings display-row index under the pointer, using the same animated
-    /// `modal_scroll_px` the rows render with — a fixed-offset hit-test drifts a row
-    /// off once the list has scrolled. `None` outside the viewport or in a row gap.
-    fn settings_row_at(&self, x: i32, y: i32, screen_w: u32, screen_h: u32) -> Option<usize> {
-        let (_, content) = view::settings::layout(&self.settings, screen_w, screen_h);
-        if !content.contains_point((x, y)) {
-            return None;
-        }
-        let stride = ui::widgets::focus_row_stride() as i32;
-        let total = menu::settings_row_count(&self.settings);
-        let scroll_px = self
-            .modal_scroll_px
-            .clamp(0, Self::max_scroll_px(total, stride, content.height()));
-        (0..total).find(|&r| ui::widgets::focus_row_rect_at_px(content, r, scroll_px).contains_point((x, y)))
-    }
-
-    /// The dropdown option index under the pointer, if a dropdown is open and the
-    /// pointer is over one of its options. Shares `dropdown_geom` +
-    /// `ui::widgets::dropdown_option_rect` with the renderer so hover previews exactly what a
-    /// click confirms.
-    fn dropdown_option_at(
-        &self,
-        x: i32,
-        y: i32,
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::text::Fonts,
-    ) -> Option<usize> {
-        let dd = self.dropdown.as_ref()?;
-        let (content, scroll_px) = self.dropdown_geom(self.screen, screen_w, screen_h, fonts)?;
-        let overlay = view::settings::dropdown_overlay_rect_at_px(content, dd.row, scroll_px);
-        let options_len = match self.screen {
-            Screen::Diagnostics => menu::LOG_LEVEL_OPTIONS.len(),
-            _ => menu::dropdown_option_count(menu::settings_logical_row(&self.settings, dd.row)),
-        };
-        (0..options_len).find(|&i| ui::widgets::dropdown_option_rect(overlay, i).contains_point((x, y)))
-    }
-
-    /// A click while a dropdown is open: an option under the pointer confirms it,
-    /// anything else dismisses (tap-outside-to-close). The hovered option is already
-    /// the cursor courtesy of `handle_mouse_motion`.
-    fn dropdown_click_event(&self, x: i32, y: i32, screen_w: u32, screen_h: u32, fonts: &ui::text::Fonts) -> MenuEvent {
-        if self.dropdown_option_at(x, y, screen_w, screen_h, fonts).is_some() {
-            MenuEvent::Confirm
-        } else {
-            MenuEvent::Back
-        }
-    }
-
-    /// The current screen's modal card rect, or `None` for a screen that draws no
-    /// modal card (Home, or Wake before its payload is set). Measured off the same
-    /// [`ui::ModalScreen`] value the renderer draws, so hover, click and the
-    /// close-button hit-test can never drift from what is on screen.
-    fn modal_card_rect(&self, screen_w: u32, screen_h: u32, fonts: &ui::text::Fonts) -> Option<Rect> {
-        self.with_modal_screen(|s| s.card_rect(screen_w, screen_h, fonts))
-    }
-
-    /// The open modal's row-list viewport, or an empty rect on a screen without one —
-    /// for the tile builders, which only reach here on a screen that has one.
-    fn modal_list_content(&self, screen_w: u32, screen_h: u32, fonts: &ui::text::Fonts) -> Rect {
-        self.modal_list_geometry(screen_w, screen_h, fonts)
-            .map_or_else(|| Rect::new(0, 0, 0, 0), |(_, content)| content)
-    }
-
-    /// The open modal's card and row-list viewport, when it has one. The pair every hit
-    /// test and focused-row tile needs, derived once from the screen that draws them.
-    fn modal_list_geometry(&self, screen_w: u32, screen_h: u32, fonts: &ui::text::Fonts) -> Option<(Rect, Rect)> {
-        self.with_modal_screen(|s| {
-            let card = s.card_rect(screen_w, screen_h, fonts);
-            s.content_rect(card, fonts).map(|content| (card, content))
-        })?
-    }
-
-    /// Calls `f` with the open modal as a [`ui::ModalScreen`], built from the state it
-    /// shows. `None` on Home, and on a screen whose payload isn't set yet (Wake before its
-    /// host is known).
-    ///
-    /// By closure rather than by return value: the hit tests below run this on every Magic
-    /// Remote `MouseMotion`, and a returned `Box<dyn ModalScreen>` would put a heap
-    /// allocation on that path for what is a geometry query.
-    fn with_modal_screen<R>(&self, f: impl FnOnce(&dyn ui::ModalScreen) -> R) -> Option<R> {
-        Some(match self.screen {
-            Screen::Home => return None,
-            Screen::Settings => f(&view::settings::Modal {
-                settings: &self.settings,
-            }),
-            Screen::Pairing => f(&view::pairing::Modal {
-                pin_digits: &self.pin_digits,
-                status: self.pairing_status.as_ref(),
-                busy: self.pairing_busy,
-            }),
-            Screen::AddHost | Screen::EditHost => {
-                let (title, subtitle) = self.address_copy();
-                f(&view::addhost::Modal {
-                    title,
-                    subtitle,
-                    typed: self.add_host.text(),
-                    keyboard_shown: self.keyboard_shown,
-                })
-            }
-            Screen::Wake => f(&view::wake::Modal {
-                wake: self.wake.as_ref()?,
-            }),
-            Screen::ForgetHost => f(&view::forget::Modal {
-                host_name: self.host_menu_host_name().unwrap_or_default(),
-            }),
-            Screen::HostMenu => f(&view::hostmenu::Modal {
-                title: self.host_menu_host_name().unwrap_or_default(),
-                subtitle: self.host_menu_subtitle(),
-                rows: self.host_menu_rows(),
-            }),
-            Screen::WakeSettings => f(&view::wakesettings::Modal {
-                host_name: self.host_menu_host_name().unwrap_or_default(),
-                auto_send: self.wake_settings_host().is_some_and(|h| h.wol_auto),
-            }),
-            Screen::About => f(&view::about::Modal),
-            Screen::SpeedTest => f(&view::speedtest::Modal {
-                state: self.speed_test.as_ref(),
-                host_name: &self.speed_test_name,
-            }),
-            Screen::PinLimit => f(&view::pinlimit::Modal {
-                message: Self::PIN_LIMIT_MESSAGE,
-            }),
-            Screen::Diagnostics => f(&view::diagnostics::Modal {
-                settings: &self.settings,
-            }),
-            Screen::Experimental => f(&view::experimental::Modal {
-                settings: &self.settings,
-                rooted: Self::rooted(),
-            }),
-            Screen::CursorSettings => f(&view::cursorsettings::Modal {
-                settings: &self.settings,
-            }),
-            Screen::SendLogs => f(&view::sendlogs::Modal),
-        })
-    }
-
-    /// A painter for the current screen's modal, sized and positioned to its *tile*
-    /// region — the card rect grown by [`MODAL_TILE_PAD`] for the shadow — rather than
-    /// to the whole screen. Records the region in `modal_tile_region`, which is where
-    /// `compose_modal` composites the tile. Falls back to full-screen on a screen with
-    /// no card (shouldn't happen with one open).
-    fn modal_painter(&mut self, screen_w: u32, screen_h: u32, fonts: &ui::text::Fonts) -> Painter {
-        let card = self.modal_card_rect(screen_w, screen_h, fonts);
-        let pad = MODAL_TILE_PAD;
-        let region = card.map_or_else(
-            || Rect::new(0, 0, screen_w, screen_h),
-            |c| {
-                Rect::new(
-                    c.x() - pad,
-                    c.y() - pad,
-                    c.width() + 2 * pad as u32,
-                    c.height() + 2 * pad as u32,
-                )
-            },
-        );
-        self.modal_tile_region = region;
-        let mut p = Painter::new(region.width(), region.height());
-        p.set_origin(region.x(), region.y());
-        p
-    }
-
-    /// The list-modal row index under the pointer, for the plain list modals whose
-    /// rows are laid out by `ui::widgets::list_modal_content_rect` + `ui::widgets::focus_row_rect`
-    /// (HostMenu/Diagnostics/Experimental). `None` on any other screen or when the
-    /// pointer misses every row. Shared so hover and click hit-test identically.
-    fn modal_list_row_at(
-        &self,
-        x: i32,
-        y: i32,
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::text::Fonts,
-    ) -> Option<usize> {
-        let card = self.modal_card_rect(screen_w, screen_h, fonts)?;
-        let (subtitle, rows) = match self.screen {
-            Screen::HostMenu => (self.host_menu_subtitle(), self.host_menu_actions().len()),
-            Screen::Diagnostics => (view::diagnostics::SUBTITLE.to_string(), menu::DIAGNOSTICS_ROW_COUNT),
-            Screen::Experimental => (
-                view::experimental::SUBTITLE.to_string(),
-                view::experimental::row_count(Self::rooted()),
-            ),
-            Screen::CursorSettings => (view::cursorsettings::SUBTITLE.to_string(), menu::CURSOR_ROW_COUNT),
-            _ => return None,
-        };
-        let content = ui::widgets::list_modal_content_rect(card, fonts, &subtitle, rows);
-        (0..rows).find(|&r| ui::widgets::focus_row_rect(content, r).contains_point((x, y)))
-    }
-
-    fn hover_close_at(&mut self, x: i32, y: i32, screen_w: u32, screen_h: u32, fonts: &ui::text::Fonts) -> bool {
-        let Some(card) = self.modal_card_rect(screen_w, screen_h, fonts) else {
-            // Home draws no close button, but `hover_close` is only ever set true by a
-            // modal branch — without clearing it on the way back to Home it stayed stuck
-            // `true` forever (nothing on Home reset it), and `handle_mouse_click`'s
-            // `if self.hover_close { return self.back() }` then swallowed every Home
-            // click. Not reported as a visible change: Home draws no close button.
-            self.hover_close = false;
-            return false;
-        };
-        self.set_hover_close(ui::widgets::modal_close_rect(card).contains_point((x, y)))
-    }
-
-    /// Updates `hover_close` and reports whether it actually changed — every modal
-    /// screen's close-button hover check in `handle_mouse_motion` follows this same
-    /// shape.
-    pub(crate) fn set_hover_close(&mut self, hover_close: bool) -> bool {
-        let changed = hover_close != self.hover_close;
-        self.hover_close = hover_close;
-        changed
-    }
-
-    /// A pointer click confirms whatever's currently hovered/focused, or triggers
-    /// Back if the modal's close (X) button itself is what's hovered.
-    pub fn handle_mouse_click(
-        &mut self,
-        x: i32,
-        y: i32,
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::text::Fonts,
-    ) -> Option<ConnectTarget> {
-        // Re-sync the close-button hover to the click's own position first — a
-        // MouseButtonDown can carry a slightly different (x, y) than the last
-        // MouseMotion (the physical button press can jostle the remote a little).
-        self.handle_mouse_motion(x, y, screen_w, screen_h, fonts);
-        if self.hover_close {
-            // Same "what Back means here" as everywhere else — see `back`'s docs.
-            return self.back();
-        }
-        // Unlike hover, a click DOES move `home_focus`/`settings_focused` — fresh at
-        // the click's own position, so it confirms what was actually clicked rather
-        // than whatever the keyboard/remote last focused elsewhere.
-        match self.screen {
-            Screen::Home => {
-                // The ⋯ button sits inside its row, so it has to be tested first or the
-                // click just reads as a click on the host.
-                if let Some(idx) =
-                    view::sidebar::hit_test_menu_button(x, y, self.entries.len(), self.sidebar_len(), screen_h)
-                {
-                    self.home_focus = HomeFocus::SidebarMenu(idx);
-                    self.open_host_menu(idx);
-                    return None;
-                }
-                if let Some(idx) = view::sidebar::hit_test_row(x, y, self.sidebar_len(), screen_h) {
-                    self.home_focus = HomeFocus::Sidebar(idx);
-                } else {
-                    let available_w = screen_w.saturating_sub(ui::widgets::SIDEBAR_W);
-                    let columns = view::home::grid_columns(available_w);
-                    // Clicked empty space — either between cards (`?`'s early
-                    // `None`) or the padding after a partial pinned row.
-                    let idx = view::home::hit_test_grid_card(
-                        x,
-                        y,
-                        columns,
-                        self.grid_len(columns),
-                        ui::widgets::SIDEBAR_W as i32,
-                        available_w,
-                        self.grid_scroll,
-                    )?;
-                    if !self.is_grid_card(idx, columns) {
-                        return None;
-                    }
-                    self.home_focus = HomeFocus::Grid(idx);
-                }
-                self.handle_home_event(MenuEvent::Confirm, screen_w, screen_h)
-            }
-            Screen::Settings => {
-                if self.dropdown.is_some() {
-                    let ev = self.dropdown_click_event(x, y, screen_w, screen_h, fonts);
-                    self.handle_settings_event(ev, screen_h);
-                    return None;
-                }
-                // `?` bails if the click hit the gap between rows or outside the
-                // viewport — nothing to focus or confirm.
-                self.settings_focused = self.settings_row_at(x, y, screen_w, screen_h)?;
-                self.handle_settings_event(MenuEvent::Confirm, screen_h);
-                None
-            }
-            Screen::Pairing => {
-                // The Magic Remote pointer is the most reliable input on this TV, so the
-                // "Request access" button is clickable directly: focus it and confirm.
-                let card = view::pairing::card_rect(screen_w, screen_h, fonts);
-                if view::pairing::request_button_rect(card, fonts).contains_point((x, y)) {
-                    self.pairing_focus = PairingFocus::RequestAccess;
-                    self.handle_pairing_event(MenuEvent::Confirm);
-                }
-                None
-            }
-            Screen::Wake => {
-                self.handle_wake_event(MenuEvent::Confirm);
-                None
-            }
-            Screen::ForgetHost => {
-                self.handle_forget_host_event(MenuEvent::Confirm);
-                None
-            }
-            // A click focuses the row it landed on first, then confirms it — same
-            // click-moves-focus rule as Home/Settings above.
-            Screen::HostMenu => {
-                let rows = self.host_menu_actions().len();
-                let (_, content) = self.modal_list_geometry(screen_w, screen_h, fonts)?;
-                let i = (0..rows).find(|&i| ui::widgets::focus_row_rect(content, i).contains_point((x, y)))?;
-                self.menu_focused = i;
-                // A click that landed on the row's ⋯ opens that instead of the row's own
-                // action — same split as a sidebar host row's button.
-                let row = ui::widgets::focus_row_rect(content, i);
-                self.host_menu_dots =
-                    self.host_menu_row_has_dots() && ui::widgets::sidebar_menu_button_rect(row).contains_point((x, y));
-                self.handle_host_menu_event(MenuEvent::Confirm);
-                None
-            }
-            Screen::WakeSettings => {
-                let (_, content) = self.modal_list_geometry(screen_w, screen_h, fonts)?;
-                if ui::widgets::focus_row_rect(content, 0).contains_point((x, y)) {
-                    self.handle_wake_settings_event(MenuEvent::Confirm);
-                }
-                None
-            }
-            Screen::SpeedTest => {
-                self.handle_speed_test_event(MenuEvent::Confirm);
-                None
-            }
-            // A click anywhere but the close button (handled above) dismisses it,
-            // same as the one OK button would — there's nothing else on this card.
-            Screen::PinLimit => {
-                self.handle_pin_limit_event(MenuEvent::Confirm);
-                None
-            }
-            Screen::Diagnostics => {
-                if self.dropdown.is_some() {
-                    let ev = self.dropdown_click_event(x, y, screen_w, screen_h, fonts);
-                    self.handle_diagnostics_event(ev);
-                    return None;
-                }
-                if let Some(row) = self.modal_list_row_at(x, y, screen_w, screen_h, fonts) {
-                    self.diagnostics_focused = row;
-                    self.handle_diagnostics_event(MenuEvent::Confirm);
-                }
-                None
-            }
-            Screen::Experimental => {
-                if let Some(row) = self.modal_list_row_at(x, y, screen_w, screen_h, fonts) {
-                    self.experimental_focused = row;
-                    self.handle_experimental_event(MenuEvent::Confirm);
-                }
-                None
-            }
-            Screen::CursorSettings => {
-                if let Some(row) = self.modal_list_row_at(x, y, screen_w, screen_h, fonts) {
-                    self.cursor_settings_focused = row;
-                    self.handle_cursor_settings_event(MenuEvent::Confirm);
-                }
-                None
-            }
-            // A click confirms whichever of Cancel/Send currently has focus —
-            // same click-confirms-the-focused-button shape as ForgetHost.
-            Screen::SendLogs => {
-                self.handle_send_logs_event(MenuEvent::Confirm);
-                None
-            }
-            // Nothing clickable but the close button (handled above).
-            Screen::AddHost | Screen::EditHost | Screen::About => None,
-        }
-    }
-    // --------------------------------------------------------------- render --
 
     /// Queues the whole document for the background writer. Every mutation of settings, hosts or
     /// selection comes through here rather than writing its own slice.
@@ -1719,1704 +1058,5 @@ impl App {
             }
         }
         screen_changed
-    }
-
-    /// Sidebar family: the focus-free strip (rebuilt on content change) plus the
-    /// single focused-row overlay tile. Pushes any rebuilt tiles onto `updated`.
-    /// Extracted from `prepare_tiles` as a self-contained family (A2 staging).
-    fn prepare_sidebar(
-        &mut self,
-        tiles: &mut TileStore,
-        text_cache: &mut crate::ui::text::TextCache,
-        fonts: &ui::text::Fonts,
-        screen_h: u32,
-        updated: &mut Vec<TileId>,
-    ) -> Result<()> {
-        // Kept on the `sidebar_dirty` flag rather than a content version: the strip is
-        // built from every entry plus its reachability, and hashing all of that once a
-        // frame would cost more than the flag the event side already maintains.
-        if self.sidebar_dirty || !tiles.contains(tile::SIDEBAR) {
-            let selected = self.sidebar_index_of_selected_host();
-            let entries = &self.entries;
-            let reach = self.reachability_list();
-            // Reuses the existing full-height strip as its own scratch surface — several MB
-            // that would otherwise be reallocated on every host list change.
-            // The outer condition has already decided this is stale, so the version just
-            // has to differ from the last one — `ensure_in_place` would otherwise see an
-            // unchanged `STATIC` and skip the rebuild it was called to do.
-            self.sidebar_gen = self.sidebar_gen.wrapping_add(1);
-            tiles.ensure_in_place(
-                tile::SIDEBAR,
-                self.sidebar_gen,
-                || Painter::new(ui::widgets::SIDEBAR_W, screen_h),
-                |layer| {
-                    view::sidebar::draw(
-                        &mut ui::Canvas {
-                            painter: layer,
-                            text_cache,
-                            fonts,
-                            // The sidebar's own strip is the full panel height and a fixed width.
-                            screen_w: ui::widgets::SIDEBAR_W,
-                            screen_h,
-                        },
-                        entries,
-                        None,
-                        selected,
-                        &reach,
-                    )
-                },
-            )?;
-            self.sidebar_dirty = false;
-            tiles.remove(tile::FOCUS_ROW); // row content may have changed under it
-            updated.push(tile::SIDEBAR);
-        }
-        // One tile serves both sidebar focus states (see `render_focused_row_tile`).
-        let sidebar_focus = match self.home_focus {
-            HomeFocus::Sidebar(i) => Some((i, false)),
-            HomeFocus::SidebarMenu(i) => Some((i, true)),
-            HomeFocus::Grid(_) => None,
-        };
-        if let Some(key) = sidebar_focus {
-            let online = self.entries.get(key.0).and_then(|e| self.entry_online(e));
-            if tiles.ensure(tile::FOCUS_ROW, cache::version(&key), || {
-                view::sidebar::render_focused_row_tile(text_cache, fonts, &self.entries, key.0, key.1, online)
-            })? {
-                updated.push(tile::FOCUS_ROW);
-            }
-        }
-        Ok(())
-    }
-
-    /// Grid family: windowed/budgeted card-tile building, eviction, the reveal
-    /// spinner, and the shared ring/outline/pin-badge tiles — or the "no host"
-    /// hint when nothing is selected. Extracted from `prepare_tiles` (A2 staging).
-    #[allow(clippy::too_many_arguments)]
-    fn prepare_grid(
-        &mut self,
-        tiles: &mut TileStore,
-        text_cache: &mut crate::ui::text::TextCache,
-        fonts: &ui::text::Fonts,
-        columns: usize,
-        card_w: u32,
-        card_h: u32,
-        screen_h: u32,
-        updated: &mut Vec<TileId>,
-    ) -> Result<()> {
-        // Reset before the branch: it is only ever set inside it, and a stale `true` left
-        // behind by a host that has since been deselected would spin the render loop at
-        // full rate forever.
-        self.tiles_pending = false;
-        if self.selected_host.is_some() {
-            let count = self.grid_len(columns);
-            // Captured before it's cleared below: a fresh library load is the only
-            // rebuild that also re-arms the spinner.
-            let full_reset = self.grid_dirty;
-            if full_reset {
-                // Every existing texture is stale (different games, different host) —
-                // drop them rather than leaving them to be overwritten one by one,
-                // which would strand the tail of a longer previous library.
-                for id in self.card_ids.release_all() {
-                    tiles.remove(id);
-                    self.evicted_tiles.push(id);
-                }
-                self.card_pop.clear();
-                self.grid_dirty = false;
-                self.grid_cards_dirty.clear();
-                // Scrolling or re-pinning a card must not hide the already-visible
-                // grid behind the spinner again.
-                self.grid_reveal_ready = false;
-                self.spinner_since = None;
-                self.spinner_frame = None;
-            } else {
-                for id in std::mem::take(&mut self.grid_cards_dirty) {
-                    if let Some(t) = self.card_ids.release(&id) {
-                        tiles.remove(t);
-                    }
-                    self.card_pop.remove(&id);
-                }
-            }
-
-            // Windowed, budgeted tile building — see `CARD_BUILD_BUDGET`.
-            let row_h = card_h as i32 + view::home::GRID_GAP;
-            let visible_rows = (screen_h as i32 - view::home::GRID_TOP_Y).max(row_h) / row_h + 1;
-            let first_visible_row = (self.grid_scroll / row_h).max(0);
-            let row_of = |idx: usize| (idx / columns.max(1)) as i32;
-            let build_lo = first_visible_row - CARD_PREFETCH_ROWS;
-            let build_hi = first_visible_row + visible_rows + CARD_PREFETCH_ROWS;
-            let keep_lo = first_visible_row - CARD_KEEP_ROWS;
-            let keep_hi = first_visible_row + visible_rows + CARD_KEEP_ROWS;
-
-            // Held by value, not re-derived per index — and, unlike the `App`
-            // helpers, it maps indices without borrowing all of `self`, so the art
-            // lookups below can sit next to `&mut self.art_loader`.
-            let layout = self.grid_layout(columns);
-
-            // Evict first, so a long scroll frees textures in the same frame it needs new
-            // ones rather than a frame later.
-            for idx in 0..count {
-                let row = row_of(idx);
-                if row < keep_lo || row > keep_hi {
-                    let Some(id) = layout.pin_id_at(&self.games, idx) else {
-                        continue;
-                    };
-                    if let Some(t) = self.card_ids.release(id) {
-                        tiles.remove(t);
-                        self.evicted_tiles.push(t);
-                    }
-                    self.card_pop.remove(id);
-                    if layout.game_at(&self.games, idx).is_some() {
-                        // Drop the decoded cover too — it is several times the size of the
-                        // tile it feeds. Re-requested from the disk cache on scroll back.
-                        self.art.remove(id);
-                        if let Some(loader) = &mut self.art_loader {
-                            loader.forget(id);
-                        }
-                    }
-                }
-            }
-
-            // Ready once nothing more can arrive: cover already in `self.art`, or the game
-            // never had one to fetch (no `self.art` entry either way). "Desktop" and the
-            // padding after a partial pinned row have no `games` entry and are always ready.
-            let art_ready = |idx: usize| {
-                layout.game_at(&self.games, idx).is_none_or(|game| {
-                    self.art.contains_key(&game.id) || (game.art.portrait.is_none() && game.art.header.is_none())
-                })
-            };
-
-            // Art-ready cards build first — building one before its cover arrives just
-            // burns a second budget slot re-dirtying it once the cover shows up.
-            let mut to_build = Vec::new();
-            for idx in 0..count {
-                let row = row_of(idx);
-                if row < build_lo || row > build_hi {
-                    continue;
-                }
-                // Nothing to build or fetch art for in the padding after a partial
-                // pinned row.
-                let Some(id) = layout.pin_id_at(&self.games, idx) else {
-                    continue;
-                };
-                // Ask for this card's cover as it enters the window, not for the whole
-                // library at once (see `art::ArtLoader`).
-                if let (Some(loader), Some(game)) = (&mut self.art_loader, layout.game_at(&self.games, idx)) {
-                    loader.request(game);
-                }
-                if self.card_ids.get(id).is_some_and(|t| tiles.contains(t)) {
-                    continue;
-                }
-                to_build.push((idx, id.to_string(), art_ready(idx)));
-            }
-            to_build.sort_by_key(|(_, _, ready)| !ready);
-
-            let mut pending = false;
-            for (built, (idx, id, _)) in to_build.into_iter().enumerate() {
-                if built >= CARD_BUILD_BUDGET {
-                    pending = true;
-                    break;
-                }
-                let tile = {
-                    let (title, art) = self.grid_card_content(idx, columns);
-                    ui::tiles::render_card_tile(text_cache, fonts, card_w, card_h, title, art)?
-                };
-                let tile_id = self.card_ids.id(&id);
-                tiles.put(tile_id, cache::STATIC, tile);
-                if self.grid_reveal_ready {
-                    self.card_pop.insert(id.clone(), Instant::now());
-                }
-                updated.push(tile_id);
-            }
-            self.tiles_pending = pending;
-
-            // Prefetch the focused card's hero, so the connecting screen has one ready the
-            // moment OK is pressed. Deduped in the loader, and the fetched bytes are
-            // disk-cached, so hovering back over a card costs no round trip.
-            //
-            // Only once the visible window has settled: the loader serves hero requests
-            // ahead of card art, so queueing one mid-scroll would put the cards the user is
-            // actually looking at behind a full-screen fetch and decode.
-            if self.grid_reveal_ready && !pending {
-                if let HomeFocus::Grid(focus_idx) = self.home_focus {
-                    if let Some(game) = layout.game_at(&self.games, focus_idx) {
-                        if let Some(loader) = &mut self.art_loader {
-                            loader.request_hero(game);
-                        }
-                        self.hero.want(&game.id);
-                    }
-                }
-            }
-
-            // The pinned badge tile — built once, composited over the focused
-            // card in `draw_list` rather than baked into individual card tiles.
-            if tiles.ensure_static(tile::PIN_BADGE, || ui::tiles::render_pin_badge_tile(text_cache, fonts))? {
-                updated.push(tile::PIN_BADGE);
-            }
-
-            // Rechecks the whole window rather than trusting `!pending`, since a card
-            // built earlier can still be waiting behind a re-dirtied sibling; requires
-            // `art_ready` too so a placeholder built this tick can't count as revealed.
-            if !self.grid_reveal_ready {
-                let window_ready = (0..count)
-                    .filter(|&idx| {
-                        let row = row_of(idx);
-                        row >= build_lo && row <= build_hi
-                    })
-                    .all(|idx| {
-                        layout
-                            .pin_id_at(&self.games, idx)
-                            .is_none_or(|id| self.card_ids.get(id).is_some_and(|t| tiles.contains(t)) && art_ready(idx))
-                    });
-                let since = *self.spinner_since.get_or_insert_with(Instant::now);
-                self.grid_reveal_ready = window_ready || since.elapsed() >= SPINNER_MAX_WAIT;
-                if self.grid_reveal_ready {
-                    self.spinner_since = None;
-                    self.spinner_frame = None;
-                    // Everything built behind the spinner becomes visible in this
-                    // one frame, so it all zooms in off a single clock.
-                    let now = Instant::now();
-                    for id in self.card_ids.pin_ids() {
-                        self.card_pop.entry(id.to_string()).or_insert(now);
-                    }
-                } else {
-                    let (frame_idx, _) = crate::assets::spinner_frame_at(since.elapsed().as_secs_f32());
-                    if self.spinner_frame != Some(frame_idx) {
-                        self.spinner_frame = Some(frame_idx);
-                        updated.push(tile::spinner(frame_idx));
-                    }
-                }
-            }
-
-            // Both are one shared tile at the current card size, so the card size *is*
-            // the version — a resolution change rebuilds them, nothing else does.
-            let size = cache::version(&(card_w, card_h));
-            if tiles.ensure(tile::RING, size, || {
-                Ok(ui::tiles::render_focus_ring_tile(card_w, card_h))
-            })? {
-                updated.push(tile::RING);
-            }
-            if tiles.ensure(tile::CARD_OUTLINE, size, || {
-                Ok(ui::tiles::render_card_outline_tile(card_w, card_h))
-            })? {
-                updated.push(tile::CARD_OUTLINE);
-            }
-        } else {
-            self.grid_reveal_ready = true;
-            self.spinner_since = None;
-            if tiles.ensure_static(tile::NO_HOST, || {
-                ui::tiles::render_text_tile(
-                    text_cache,
-                    fonts,
-                    fonts.label,
-                    "No host selected — pick one from the list, or add one.",
-                    ui::style::theme().muted,
-                )
-            })? {
-                updated.push(tile::NO_HOST);
-            }
-        }
-        Ok(())
-    }
-
-    /// Uploads the launching game's hero art as `tile::HERO`, once, and starts its
-    /// fade-in clock. Gated on the launch having actually started: at ~1600px wide this
-    /// is a multi-MB texture, and putting one on the GPU for every card the user merely
-    /// scrolls past would undo the whole point of the windowed card cache.
-    fn prepare_hero(&mut self, updated: &mut Vec<TileId>) {
-        if self.launch_anim.is_none() {
-            return;
-        }
-        let Some(id) = self.hero.pending_upload() else { return };
-        // One hero slot, so replacing one means dropping the old texture first —
-        // `Compositor::upload_raw` treats an existing texture as already correct.
-        if self.hero.mark_uploaded(id).is_some() {
-            self.evicted_tiles.push(tile::HERO);
-        }
-        updated.push(tile::HERO);
-    }
-
-    /// Modal family: the open modal's full-screen shell tile (rebuilt only on content
-    /// change, keyed by `ModalShellKey`) and its single focused-widget tile (keyed by
-    /// `ModalFocusKey`). Extracted from `prepare_tiles` (A2 staging).
-    #[allow(clippy::too_many_arguments)]
-    fn prepare_modal(
-        &mut self,
-        tiles: &mut TileStore,
-        text_cache: &mut crate::ui::text::TextCache,
-        fonts: &ui::text::Fonts,
-        screen_w: u32,
-        screen_h: u32,
-        content_dirty: bool,
-        screen_changed: bool,
-        updated: &mut Vec<TileId>,
-    ) -> Result<()> {
-        let modal_open = !matches!(self.screen, Screen::Home);
-        // Every modal's shell only reacts to *content* changes — not to
-        // `content_dirty`, which is also `true` on plain focus movement (see
-        // `ModalShellKey`'s docs). `AddHost` has no `ModalShellKey` variant
-        // (no split focus tile to protect) and just redraws on any
-        // `content_dirty` tick, same as every modal did before this split.
-        let modal_shell_key = match self.screen {
-            Screen::Settings => Some(ModalShellKey::Settings {
-                show_bitrate_warning: self.settings.bitrate_kbps > menu::BITRATE_WARN_KBPS,
-                hover_close: self.hover_close,
-            }),
-            Screen::Wake => self.wake.as_ref().map(|w| ModalShellKey::Wake {
-                name: w.name.clone(),
-                mac_empty: w.mac.is_empty(),
-                sent: w.sent,
-                hover_close: self.hover_close,
-            }),
-            Screen::Pairing => Some(ModalShellKey::Pairing {
-                digits: self.pin_digits,
-                status: self.pairing_status.clone(),
-                busy: self.pairing_busy,
-                hover_close: self.hover_close,
-            }),
-            Screen::ForgetHost => Some(ModalShellKey::ForgetHost {
-                name: self
-                    .host_menu_index
-                    .and_then(|i| self.entries.get(i))
-                    .map(|e| e.name().to_string()),
-                hover_close: self.hover_close,
-            }),
-            Screen::HostMenu => Some(ModalShellKey::HostMenu {
-                name: self.host_menu_title(),
-                subtitle: self.host_menu_subtitle(),
-                rows: self.host_menu_actions().len(),
-                hover_close: self.hover_close,
-            }),
-            Screen::WakeSettings => Some(ModalShellKey::WakeSettings {
-                title: view::wakesettings::title(&self.host_menu_title()),
-                auto: self.wake_settings_host().is_some_and(|h| h.wol_auto),
-                hover_close: self.hover_close,
-            }),
-            Screen::About => Some(ModalShellKey::About {
-                hover_close: self.hover_close,
-            }),
-            // The whole shell is derived from the status sentence, which already encodes
-            // the phase and the latest measurement.
-            Screen::SpeedTest => Some(ModalShellKey::SpeedTest {
-                status: view::speedtest::status(self.speed_test.as_ref(), &self.speed_test_name),
-                hover_close: self.hover_close,
-            }),
-            Screen::Diagnostics => Some(ModalShellKey::Diagnostics {
-                log_level: self.settings.log_level_override,
-                stats_overlay: self.settings.stats_overlay,
-                show_logs: self.settings.show_logs,
-                hover_close: self.hover_close,
-            }),
-            Screen::Experimental => Some(ModalShellKey::Experimental {
-                video_pacing: self.settings.video_pacing,
-                game_mode: self.settings.game_mode,
-                hover_close: self.hover_close,
-            }),
-            Screen::CursorSettings => Some(ModalShellKey::CursorSettings {
-                cursor_capture: self.settings.cursor_capture,
-                cursor_gestures: self.settings.cursor_gestures,
-                hover_close: self.hover_close,
-            }),
-            Screen::SendLogs => Some(ModalShellKey::SendLogs {
-                hover_close: self.hover_close,
-            }),
-            // `EditHost` joins `AddHost` in having no shell key: its typed-digit
-            // display has no separate focus tile to protect, so it just redraws on
-            // any `content_dirty` tick — same for `PinLimit`, which is a fixed
-            // message plus one always-focused button.
-            Screen::Home | Screen::AddHost | Screen::EditHost | Screen::PinLimit => None,
-        };
-        let modal_stale = if modal_shell_key.is_some() {
-            !tiles.contains(tile::MODAL) || self.modal_shell_key != modal_shell_key
-        } else {
-            content_dirty || !tiles.contains(tile::MODAL)
-        };
-        self.modal_shell_key = modal_shell_key;
-        if modal_open && (screen_changed || modal_stale) {
-            // Sized to the card's bounding box, not the whole screen: the render fns
-            // below draw at absolute, screen-centered coordinates, and the painter's
-            // origin shift (see `Painter::set_origin`) maps that geometry into the
-            // smaller buffer.
-            let mut p = self.modal_painter(screen_w, screen_h, fonts);
-            let c = &mut ui::Canvas::new(&mut p, text_cache, fonts, screen_w, screen_h);
-            let hover_close = self.hover_close;
-            self.with_modal_screen(|s| s.render(c, hover_close)).transpose()?;
-            // Staleness is decided above (the shell key is compared against the previous
-            // frame's, not hashed), so the store just takes the result.
-            tiles.put(tile::MODAL, cache::STATIC, p);
-            updated.push(tile::MODAL);
-        }
-        // Whichever modal is open has at most one focused, zoom-animated widget
-        // (`ModalFocusKey`'s docs) — `None` for screens with no such widget
-        // (Home, AddHost) or when Wake has nothing to focus (no MAC on record,
-        // see `handle_wake_event`'s matching guard).
-        let focus_key = match self.screen {
-            Screen::Settings => Some(ModalFocusKey::SettingsRow(
-                self.settings_focused,
-                self.settings,
-                self.detected_gamepad_type,
-            )),
-            Screen::Wake => self
-                .wake
-                .as_ref()
-                .filter(|w| !w.mac.is_empty())
-                .map(|w| ModalFocusKey::WakeButton(w.focused)),
-            Screen::Pairing => Some(match self.pairing_focus {
-                PairingFocus::Pin => {
-                    ModalFocusKey::PairingDigit(self.pin_digit_index, self.pin_digits[self.pin_digit_index])
-                }
-                PairingFocus::RequestAccess => ModalFocusKey::PairingButton,
-            }),
-            Screen::ForgetHost => Some(ModalFocusKey::ForgetButton(self.host_menu_focused)),
-            Screen::HostMenu => self
-                .host_menu_actions()
-                .get(self.menu_focused)
-                .map(|(_, row)| ModalFocusKey::MenuRow(self.menu_focused, row.label.clone(), self.host_menu_dots)),
-            Screen::WakeSettings => Some(ModalFocusKey::WakeToggle(
-                self.wake_settings_host().is_some_and(|h| h.wol_auto),
-            )),
-            // Only once there are buttons to focus — while measuring there is nothing
-            // on the card but text.
-            Screen::SpeedTest => view::speedtest::finished(self.speed_test.as_ref()).then(|| {
-                let label = view::speedtest::apply_label(view::speedtest::recommendation(self.speed_test.as_ref()));
-                ModalFocusKey::SpeedTestButton(self.speed_test_focused, label)
-            }),
-            Screen::Diagnostics => Some(ModalFocusKey::DiagnosticsRow(
-                self.diagnostics_focused,
-                self.settings.log_level_override,
-                self.settings.stats_overlay,
-                self.settings.show_logs,
-            )),
-            Screen::Experimental => Some(ModalFocusKey::ExperimentalRow(
-                self.experimental_focused,
-                self.settings.video_pacing,
-                self.settings.game_mode,
-            )),
-            Screen::CursorSettings => Some(ModalFocusKey::CursorSettingsRow(
-                self.cursor_settings_focused,
-                self.settings.cursor_capture,
-                self.settings.cursor_gestures,
-            )),
-            Screen::SendLogs => Some(ModalFocusKey::SendLogsButton(self.send_logs_focused)),
-            // Neither has a single focused widget: the address form is one always-active
-            // field, About is a scrolling document, and `PinLimit`'s one button is
-            // always drawn focused directly in `render_pin_limit`.
-            Screen::Home | Screen::AddHost | Screen::EditHost | Screen::About | Screen::PinLimit => None,
-        };
-        if let Some(key) = focus_key {
-            // Also stale on every tick of an in-flight `switch_anim`: the knob's
-            // position depends on elapsed time, not on `key`, which doesn't
-            // change mid-flip.
-            let stale = self.switch_anim.is_some() || !tiles.is_fresh(tile::MODAL_FOCUS, cache::version(&key));
-            if stale {
-                let tile = match self.screen {
-                    Screen::Settings => {
-                        let (_, content) = view::settings::layout(&self.settings, screen_w, screen_h);
-                        let rows = self.settings_rows();
-                        let dropdown_open = self.dropdown.as_ref().is_some_and(|dd| dd.row == self.settings_focused);
-                        let target_on = rows.get(self.settings_focused).is_some_and(|r| r.value == "On");
-                        ui::widgets::render_focus_row_tile(
-                            text_cache,
-                            fonts,
-                            &rows,
-                            content.width(),
-                            self.settings_focused,
-                            dropdown_open,
-                            self.toggle_frac(target_on, self.settings_focused),
-                        )?
-                    }
-                    Screen::Wake => {
-                        let wake = self
-                            .wake
-                            .as_ref()
-                            .expect("focus_key only Some for a Wake with a focusable widget");
-                        // focus_key is only Some for a Wake with a MAC, so this is the confirm dialog.
-                        let rect = Self::confirm_focus_button_rect(
-                            screen_w,
-                            screen_h,
-                            fonts,
-                            &view::wake::status_text(wake),
-                            wake.focused,
-                        );
-                        let buttons = view::wake::buttons();
-                        ui::widgets::render_confirm_button_tile(
-                            text_cache,
-                            fonts,
-                            &buttons[wake.focused],
-                            rect.width(),
-                            rect.height(),
-                        )?
-                    }
-                    Screen::Pairing => match self.pairing_focus {
-                        PairingFocus::Pin => {
-                            view::pairing::render_digit_tile(text_cache, fonts, self.pin_digits[self.pin_digit_index])?
-                        }
-                        PairingFocus::RequestAccess => {
-                            let card = view::pairing::card_rect(screen_w, screen_h, fonts);
-                            let btn = view::pairing::request_button_rect(card, fonts);
-                            view::pairing::render_button_tile(text_cache, fonts, btn.width(), btn.height())?
-                        }
-                    },
-                    Screen::ForgetHost => {
-                        let name = self
-                            .host_menu_index
-                            .and_then(|i| self.entries.get(i))
-                            .map(HostEntry::name)
-                            .unwrap_or_default();
-                        let rect = Self::confirm_focus_button_rect(
-                            screen_w,
-                            screen_h,
-                            fonts,
-                            &view::forget::subtitle(name),
-                            self.host_menu_focused,
-                        );
-                        let buttons = view::forget::buttons();
-                        ui::widgets::render_confirm_button_tile(
-                            text_cache,
-                            fonts,
-                            &buttons[self.host_menu_focused],
-                            rect.width(),
-                            rect.height(),
-                        )?
-                    }
-                    Screen::HostMenu => {
-                        let mut rows = self.host_menu_rows();
-                        // The only place a row's ⋯ is drawn lit — see `host_menu_actions`.
-                        if let Some(row) = rows.get_mut(self.menu_focused) {
-                            row.menu = row.menu.map(|_| self.host_menu_dots);
-                        }
-                        let content = self.modal_list_content(screen_w, screen_h, fonts);
-                        ui::widgets::render_focus_row_tile(
-                            text_cache,
-                            fonts,
-                            &rows,
-                            content.width(),
-                            self.menu_focused,
-                            false,
-                            0.0,
-                        )?
-                    }
-                    Screen::WakeSettings => {
-                        let on = self.wake_settings_host().is_some_and(|h| h.wol_auto);
-                        let rows = view::wakesettings::rows(on);
-                        let content = self.modal_list_content(screen_w, screen_h, fonts);
-                        ui::widgets::render_focus_row_tile(
-                            text_cache,
-                            fonts,
-                            &rows,
-                            content.width(),
-                            self.wake_settings_focused,
-                            false,
-                            self.toggle_frac(on, self.wake_settings_focused),
-                        )?
-                    }
-                    Screen::SpeedTest => {
-                        let card = view::speedtest::card_rect(
-                            screen_w,
-                            screen_h,
-                            fonts,
-                            self.speed_test.as_ref(),
-                            &self.speed_test_name,
-                        );
-                        let rect = ui::widgets::confirm_button_rect(
-                            view::speedtest::buttons_rect(card, fonts, self.speed_test.as_ref(), &self.speed_test_name),
-                            self.speed_test_focused,
-                        );
-                        let apply_label =
-                            view::speedtest::apply_label(view::speedtest::recommendation(self.speed_test.as_ref()));
-                        let buttons = view::speedtest::buttons(&apply_label);
-                        ui::widgets::render_confirm_button_tile(
-                            text_cache,
-                            fonts,
-                            &buttons[self.speed_test_focused],
-                            rect.width(),
-                            rect.height(),
-                        )?
-                    }
-                    Screen::Diagnostics => {
-                        let rows = view::diagnostics::rows(&self.settings);
-                        let content = self.modal_list_content(screen_w, screen_h, fonts);
-                        let dropdown_open = self
-                            .dropdown
-                            .as_ref()
-                            .is_some_and(|dd| dd.row == self.diagnostics_focused);
-                        let target_on = rows.get(self.diagnostics_focused).is_some_and(|r| r.value == "On");
-                        ui::widgets::render_focus_row_tile(
-                            text_cache,
-                            fonts,
-                            &rows,
-                            content.width(),
-                            self.diagnostics_focused,
-                            dropdown_open,
-                            self.toggle_frac(target_on, self.diagnostics_focused),
-                        )?
-                    }
-                    Screen::Experimental => {
-                        let rows = view::experimental::rows(&self.settings, Self::rooted());
-                        let content = self.modal_list_content(screen_w, screen_h, fonts);
-                        let target_on = rows.get(self.experimental_focused).is_some_and(|r| r.value == "On");
-                        ui::widgets::render_focus_row_tile(
-                            text_cache,
-                            fonts,
-                            &rows,
-                            content.width(),
-                            self.experimental_focused,
-                            false,
-                            self.toggle_frac(target_on, self.experimental_focused),
-                        )?
-                    }
-                    Screen::CursorSettings => {
-                        let rows = view::cursorsettings::rows(&self.settings);
-                        let content = self.modal_list_content(screen_w, screen_h, fonts);
-                        let target_on = rows.get(self.cursor_settings_focused).is_some_and(|r| r.value == "On");
-                        ui::widgets::render_focus_row_tile(
-                            text_cache,
-                            fonts,
-                            &rows,
-                            content.width(),
-                            self.cursor_settings_focused,
-                            false,
-                            self.toggle_frac(target_on, self.cursor_settings_focused),
-                        )?
-                    }
-                    Screen::SendLogs => {
-                        let rect = Self::confirm_focus_button_rect(
-                            screen_w,
-                            screen_h,
-                            fonts,
-                            view::sendlogs::SUBTITLE,
-                            self.send_logs_focused,
-                        );
-                        let buttons = view::sendlogs::buttons();
-                        ui::widgets::render_confirm_button_tile(
-                            text_cache,
-                            fonts,
-                            &buttons[self.send_logs_focused],
-                            rect.width(),
-                            rect.height(),
-                        )?
-                    }
-                    Screen::Home | Screen::AddHost | Screen::EditHost | Screen::About | Screen::PinLimit => {
-                        unreachable!("focus_key checked above")
-                    }
-                };
-                tiles.put(tile::MODAL_FOCUS, cache::version(&key), tile);
-                updated.push(tile::MODAL_FOCUS);
-            }
-        } else {
-            tiles.remove(tile::MODAL_FOCUS);
-        }
-        Ok(())
-    }
-
-    /// Dropdown family: the overlay panel + focused-option tile for an open Settings/Diagnostics dropdown; cleared when closed (unless a close-fade still needs them). Extracted from `prepare_tiles`.
-    fn prepare_dropdown(
-        &mut self,
-        tiles: &mut TileStore,
-        text_cache: &mut crate::ui::text::TextCache,
-        fonts: &ui::text::Fonts,
-        screen_w: u32,
-        screen_h: u32,
-        updated: &mut Vec<TileId>,
-    ) -> Result<()> {
-        if let Some(dd) = &self.dropdown {
-            let (options, content_w) = match self.screen {
-                Screen::Diagnostics => {
-                    let content = self.modal_list_content(screen_w, screen_h, fonts);
-                    (menu::log_level_dropdown_options(), content.width())
-                }
-                _ => {
-                    let (_, content) = view::settings::layout(&self.settings, screen_w, screen_h);
-                    let logical = menu::settings_logical_row(&self.settings, dd.row);
-                    (
-                        menu::dropdown_options(logical, self.detected_gamepad_type),
-                        content.width(),
-                    )
-                }
-            };
-
-            // Keyed by screen as well as row: row 0 means a different setting on Settings
-            // than it does on Diagnostics.
-            let overlay = cache::version(&(self.screen, dd.row));
-            if tiles.ensure(tile::DROPDOWN_OVERLAY, overlay, || {
-                let overlay_h = options.len() as u32 * ui::widgets::DROPDOWN_OPTION_H;
-                let mut p = Painter::new(content_w, overlay_h.max(1));
-                let rect = Rect::new(0, 0, content_w, overlay_h);
-                ui::Canvas::tile(&mut p, text_cache, fonts)
-                    .render(ui::widgets::DropdownOverlay::new(&options), rect)?;
-                Ok(p)
-            })? {
-                updated.push(tile::DROPDOWN_OVERLAY);
-            }
-
-            let focused = cache::version(&(self.screen, dd.row, dd.focused));
-            if tiles.ensure(tile::DROPDOWN_FOCUS, focused, || {
-                let option = options.get(dd.focused).map_or("", String::as_str);
-                ui::widgets::render_dropdown_option_tile(text_cache, fonts, option, content_w)
-            })? {
-                updated.push(tile::DROPDOWN_FOCUS);
-            }
-        } else if self.dropdown_fade.closing_frame(DROPDOWN_FADE).is_none() {
-            // Keep the tiles cached while a close-fade is in flight — `draw_list`
-            // still composites them at falling alpha.
-            tiles.remove(tile::DROPDOWN_OVERLAY);
-            tiles.remove(tile::DROPDOWN_FOCUS);
-        }
-        Ok(())
-    }
-
-    /// Scroll family: the indicator, edge-fade ramps, and windowed content tile for whichever modal overflows (Settings rows / About document). Extracted from `prepare_tiles`.
-    fn prepare_scroll(
-        &mut self,
-        tiles: &mut TileStore,
-        text_cache: &mut crate::ui::text::TextCache,
-        fonts: &ui::text::Fonts,
-        screen_w: u32,
-        screen_h: u32,
-        updated: &mut Vec<TileId>,
-    ) -> Result<()> {
-        // Whichever modal's content overflows its viewport (Settings' rows, About's
-        // document) gets its scroll indicator and content tile refreshed here — see
-        // `scroll_geometry`'s docs for why this one block covers every such modal
-        // instead of being hand-copied per screen.
-        if matches!(self.screen, Screen::About) {
-            // Mutates `about_wrapped` only — must happen before `scroll_geometry`
-            // (a `&self` read) can report a non-zero total for this frame.
-            let card = view::about::card_rect(screen_w, screen_h);
-            let body = view::about::body_rect(card, fonts);
-            self.ensure_about_wrapped(fonts, body.width());
-        }
-        if let Some((total, visible, _, content)) = self.scroll_geometry(screen_w, screen_h, fonts) {
-            let scroll = self.scroll.clamped(total, visible);
-            let ind = cache::version(&(self.screen, total, visible, scroll, content.height()));
-            if tiles.ensure(tile::SCROLL_INDICATOR, ind, || {
-                Ok(ui::widgets::render_list_scrollbar_tile(
-                    SCROLL_INDICATOR_TILE_W,
-                    content.height(),
-                    total,
-                    visible,
-                    scroll,
-                ))
-            })? {
-                updated.push(tile::SCROLL_INDICATOR);
-            }
-            // Static ramps, so these are once-per-run bakes rather than keyed rebuilds —
-            // scrolling and resizing both leave them valid (the GPU restretches them).
-            if tiles.ensure_static(tile::SCROLL_FADE, || {
-                Ok(ui::widgets::render_scroll_fade_tile(ui::widgets::FadeEdge::Bottom))
-            })? {
-                updated.push(tile::SCROLL_FADE);
-            }
-            if tiles.ensure_static(tile::SCROLL_FADE_TOP, || {
-                Ok(ui::widgets::render_scroll_fade_tile(ui::widgets::FadeEdge::Top))
-            })? {
-                updated.push(tile::SCROLL_FADE_TOP);
-            }
-            let stride = self.scroll_stride(fonts);
-            self.sync_modal_scroll(self.screen, total, visible, content.height(), stride);
-
-            match self.screen {
-                Screen::Settings => {
-                    let dropdown_row = self.dropdown.as_ref().map(|dd| dd.row);
-                    let key = cache::version(&(
-                        Screen::Settings,
-                        ScrollContentKey::Settings(self.settings, dropdown_row, self.detected_gamepad_type),
-                    ));
-                    if !tiles.is_fresh(tile::SCROLL_CONTENT, key) {
-                        let rows = self.settings_rows();
-                        let tile = ui::widgets::render_focus_rows_tile(
-                            text_cache,
-                            fonts,
-                            &rows,
-                            content.width(),
-                            dropdown_row,
-                        )?;
-                        tiles.put(tile::SCROLL_CONTENT, key, tile);
-                        // Settings' whole row list always fits one tile — no windowing.
-                        self.content_window = ui::scroll::ContentWindow {
-                            start: 0,
-                            len: menu::settings_row_count(&self.settings),
-                        };
-                        updated.push(tile::SCROLL_CONTENT);
-                    }
-                }
-                Screen::About => {
-                    if let Some(new_start) = self.content_window.recenter_if_needed(
-                        scroll,
-                        visible,
-                        total,
-                        ABOUT_WINDOW_BUDGET,
-                        ABOUT_WINDOW_MARGIN,
-                    ) {
-                        let len = ABOUT_WINDOW_BUDGET.min(total.saturating_sub(new_start));
-                        if let Some((_, wrapped)) = &self.about_wrapped {
-                            let stride = self.scroll_stride(fonts) as u32;
-                            let mut p = Painter::new(content.width().max(1), (len as u32 * stride).max(1));
-                            let mut c = ui::Canvas::tile(&mut p, text_cache, fonts);
-                            view::about::draw_window(&mut c, fonts.value, wrapped, new_start, len)?;
-                            self.content_window = ui::scroll::ContentWindow { start: new_start, len };
-                            let key = cache::version(&(Screen::About, ScrollContentKey::About(new_start)));
-                            tiles.put(tile::SCROLL_CONTENT, key, p);
-                            updated.push(tile::SCROLL_CONTENT);
-                        }
-                    }
-                }
-                _ => {}
-            }
-        }
-        Ok(())
-    }
-
-    /// Warms the caches a modal's first open would otherwise fill cold, off the
-    /// open path — call once on an idle Home frame. The first real open pays a
-    /// burst of `SDL2_ttf` glyph renders plus the card's shadow blur; on the very
-    /// first open of a session nothing is cached, which is the hitch the user
-    /// sees. Rendering the Settings shell + rows into throwaway painters here has
-    /// three surviving side effects: `text_cache` (per-line pixmaps), the
-    /// thread-local shadow/glow caches, and `SDL2_ttf`'s own freetype glyph cache.
-    /// The last is font-wide, so it speeds up every modal's cold renders — not
-    /// just Settings — even where the exact strings differ (e.g. the host menu).
-    pub fn prewarm_modal_caches(
-        &self,
-        text_cache: &mut crate::ui::text::TextCache,
-        fonts: &ui::text::Fonts,
-        screen_w: u32,
-        screen_h: u32,
-    ) -> Result<()> {
-        let mut scratch = Painter::new(screen_w, screen_h);
-        view::settings::render(
-            &mut ui::Canvas::new(&mut scratch, text_cache, fonts, screen_w, screen_h),
-            &self.settings,
-            self.hover_close,
-        )?;
-        let (_, content) = view::settings::layout(&self.settings, screen_w, screen_h);
-        let rows = self.settings_rows();
-        let _ = ui::widgets::render_focus_rows_tile(text_cache, fonts, &rows, content.width(), None)?;
-        let _ = ui::widgets::render_focus_row_tile(text_cache, fonts, &rows, content.width(), 0, false, 0.0)?;
-        Ok(())
-    }
-
-    /// Rasterizes every stale tile (tiny-skia, CPU — the only place rasterization
-    /// happens) and returns which tiles need their GPU texture re-uploaded.
-    /// `content_dirty` is the main loop's "an event/drain changed something this
-    /// tick" flag — it forces the open modal's tile to re-rasterize, since modal
-    /// content has no finer dirty tracking of its own. Pure animation frames pass
-    /// `false` and rasterize nothing at all. Call `advance_frame` first.
-    #[allow(clippy::too_many_arguments)]
-    pub fn prepare_tiles(
-        &mut self,
-        tiles: &mut TileStore,
-        text_cache: &mut crate::ui::text::TextCache,
-        fonts: &ui::text::Fonts,
-        screen_w: u32,
-        screen_h: u32,
-        content_dirty: bool,
-        screen_changed: bool,
-    ) -> Result<Vec<TileId>> {
-        let mut updated = Vec::new();
-        let available_w = screen_w.saturating_sub(ui::widgets::SIDEBAR_W);
-        let columns = view::home::grid_columns(available_w);
-        // `self.card_size` is set by `advance_frame` (same formula) before this runs; the
-        // local copy is what the tile-build loop below reads.
-        let (card_w, card_h) = view::home::grid_card_size(available_w, columns);
-
-        self.prepare_sidebar(tiles, text_cache, fonts, screen_h, &mut updated)?;
-
-        self.prepare_grid(
-            tiles,
-            text_cache,
-            fonts,
-            columns,
-            card_w,
-            card_h,
-            screen_h,
-            &mut updated,
-        )?;
-
-        self.prepare_hero(&mut updated);
-
-        // Status line block — built whenever `home_status` is set, independent of
-        // whether a host is selected (the "Send logs" result shows here too).
-        match &self.home_status {
-            Some(s) => {
-                if tiles.ensure(tile::STATUS, cache::version(s), || {
-                    let avail = screen_w.saturating_sub(ui::widgets::SIDEBAR_W);
-                    let max_w = avail.saturating_sub(2 * view::home::GRID_PAD as u32);
-                    ui::tiles::render_wrapped_text_tile(
-                        text_cache,
-                        fonts,
-                        fonts.label,
-                        s,
-                        max_w,
-                        ui::style::theme().muted,
-                        6,
-                    )
-                })? {
-                    updated.push(tile::STATUS);
-                }
-            }
-            None => {
-                tiles.remove(tile::STATUS);
-            }
-        }
-
-        self.prepare_modal(
-            tiles,
-            text_cache,
-            fonts,
-            screen_w,
-            screen_h,
-            content_dirty,
-            screen_changed,
-            &mut updated,
-        )?;
-
-        self.prepare_dropdown(tiles, text_cache, fonts, screen_w, screen_h, &mut updated)?;
-
-        self.prepare_scroll(tiles, text_cache, fonts, screen_w, screen_h, &mut updated)?;
-        Ok(updated)
-    }
-
-    /// `(total units, visible units, card rect, content/viewport rect)` for whichever
-    /// scrollable modal is open — `None` if `self.screen` has no overflowing content.
-    /// The one place this per-modal geometry lives, shared by `prepare_tiles`'s
-    /// staleness checks and `draw_list`'s GPU-crop math so the two can't disagree.
-    /// `About`'s `total` depends on `about_wrapped` already being fresh for this
-    /// frame's body width — `prepare_tiles` ensures that before calling this;
-    /// `draw_list` runs after `prepare_tiles` in the same frame, so it's already set.
-    pub(crate) fn scroll_geometry(
-        &self,
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::text::Fonts,
-    ) -> Option<(usize, usize, Rect, Rect)> {
-        self.scroll_geometry_for(self.screen, screen_w, screen_h, fonts)
-    }
-
-    /// Same as `scroll_geometry`, but for an explicit screen rather than
-    /// `self.screen` — `draw_list`'s closing-fade needs the screen it captured at
-    /// `back()` time, not whatever `self.screen` (already `Home`) says now.
-    pub(crate) fn scroll_geometry_for(
-        &self,
-        screen: Screen,
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::text::Fonts,
-    ) -> Option<(usize, usize, Rect, Rect)> {
-        match screen {
-            Screen::Settings => {
-                let (card, content) = view::settings::layout(&self.settings, screen_w, screen_h);
-                let visible = view::settings::visible_rows(&self.settings, screen_h);
-                Some((menu::settings_row_count(&self.settings), visible, card, content))
-            }
-            Screen::About => {
-                let card = view::about::card_rect(screen_w, screen_h);
-                let body = view::about::body_rect(card, fonts);
-                let total = self.about_wrapped.as_ref().map_or(0, |(_, v)| v.len());
-                let visible = view::about::visible_lines(body, fonts.raster, fonts.value);
-                Some((total, visible, card, body))
-            }
-            _ => None,
-        }
-    }
-
-    /// Clips a tile's destination to `clip`, returning `(source crop, clipped destination)`.
-    ///
-    /// The tile's full extent is assumed to map onto `dst`, so the crop is proportional —
-    /// which keeps this correct even while `dst` is being zoom-animated. `None` when nothing
-    /// of it remains inside.
-    fn clip_tile(dst: Rect, clip: Rect, tile_w: u32, tile_h: u32) -> Option<(Rect, Rect)> {
-        let visible = dst.intersection(clip)?;
-        if visible == dst {
-            return Some((Rect::new(0, 0, tile_w, tile_h), dst));
-        }
-        if dst.width() == 0 || dst.height() == 0 {
-            return None;
-        }
-        let fx = |v: i32| (f64::from(v) / f64::from(dst.width())) * f64::from(tile_w);
-        let fy = |v: i32| (f64::from(v) / f64::from(dst.height())) * f64::from(tile_h);
-        let src = Rect::new(
-            fx(visible.x() - dst.x()).round() as i32,
-            fy(visible.y() - dst.y()).round() as i32,
-            (fx(visible.width() as i32).round() as u32).max(1),
-            (fy(visible.height() as i32).round() as u32).max(1),
-        );
-        Some((src, visible))
-    }
-
-    /// The furthest the viewport may be cropped down: the last unit sits flush with the
-    /// viewport's bottom edge rather than scrolling past it.
-    ///
-    /// This is why the rendered offset is pixels and not units — `offset * stride` overshoots
-    /// by exactly the peek strip at the end of the list, which would show a dead band below
-    /// the final row (and is what the row-quantized version did).
-    fn max_scroll_px(total: usize, stride: i32, viewport_h: u32) -> i32 {
-        (total as i32 * stride - viewport_h as i32).max(0)
-    }
-
-    /// Re-derives `modal_scroll_target_px` from the integral offset, snapping rather than
-    /// gliding when the scrolling modal changed. Called once per frame from `update_tiles`,
-    /// which is where the geometry (and the fonts About's stride needs) is already in hand.
-    ///
-    /// Kept in absolute content pixels, *not* relative to the baked window: About re-bakes its
-    /// window later in the same pass, and a window-relative target would jump by the whole
-    /// window offset on the frame that happens — a full-document glide instead of a scroll.
-    /// `draw_list` subtracts the window when it crops.
-    fn sync_modal_scroll(&mut self, screen: Screen, total: usize, visible: usize, viewport_h: u32, stride: i32) {
-        let offset = self.scroll.clamped(total, visible);
-        // Biased back by one peek so the *top* edge also cuts mid-row: sitting on the row grid
-        // would put nothing but the gap between rows under the top fade, which is invisible
-        // (see `view::settings::PEEK`). The clamps then pin the first and last positions flush,
-        // where there is genuinely nothing beyond the edge to hint at.
-        let bias = match screen {
-            Screen::Settings => view::settings::PEEK as i32,
-            _ => 0,
-        };
-        let target = (offset as i32 * stride - bias)
-            .min(Self::max_scroll_px(total, stride, viewport_h))
-            .max(0);
-        self.modal_scroll_target_px = target;
-        if self.modal_scroll_screen != Some(screen) {
-            self.modal_scroll_screen = Some(screen);
-            self.modal_scroll_px = target;
-        }
-    }
-
-    /// Pixel stride between two consecutive units of whichever modal is scrolling —
-    /// Settings' fixed row height, or About's wrapped-line height. Only meaningful
-    /// when `scroll_geometry` returns `Some`.
-    fn scroll_stride(&self, fonts: &ui::text::Fonts) -> i32 {
-        self.scroll_stride_for(self.screen, fonts)
-    }
-
-    /// Same as `scroll_stride`, but for an explicit screen — see `scroll_geometry_for`.
-    fn scroll_stride_for(&self, screen: Screen, fonts: &ui::text::Fonts) -> i32 {
-        match screen {
-            Screen::Settings => ui::widgets::FOCUS_ROW_H as i32 + ui::widgets::FOCUS_ROW_GAP,
-            Screen::About => view::about::line_stride(fonts.raster, fonts.value),
-            _ => 1,
-        }
-    }
-
-    /// Builds this frame's draw list (paint order) from the current state and
-    /// animation clocks — pure bookkeeping, no rasterization (the font
-    /// params are only for pure geometry — `ui::text::modal_header_end_y` and
-    /// friends — needed to position a modal's focused-widget tile without
-    /// re-rendering its header). The GPU executes it (`Compositor::execute`).
-    /// Assembles the read-only view of state the render path consumes (see
-    /// `render_input::RenderInput`). Grows as families migrate off direct `self` reads.
-    pub fn render_input(&self) -> render_input::RenderInput<'_> {
-        render_input::RenderInput {
-            home_focus: self.home_focus,
-            entries: &self.entries,
-            host_selected: self.selected_host.is_some(),
-            has_status: self.home_status.is_some(),
-            grid_reveal_ready: self.grid_reveal_ready,
-        }
-    }
-
-    /// Modal family compose: the fade-in scrim + shell, scrollable content crop with
-    /// its edge fades, the dropdown overlay, the focused-widget zoom, and the scroll
-    /// indicator — all driven by the modal fade clock. Extracted from `draw_list`.
-    fn compose_modal(
-        &self,
-        tiles: &TileStore,
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::text::Fonts,
-        cmds: &mut Vec<DrawCmd>,
-    ) {
-        // While closing, `self.screen` has already moved on — render the fade's
-        // captured screen instead, so the still-uploaded tiles keep drawing for one
-        // more `MODAL_FADE` with alpha running in reverse (see `ui::fade::ModalFade`).
-        let closing_frame = self.modal_fade.closing_frame(MODAL_FADE);
-        let (screen, m) = match closing_frame {
-            Some((alpha, s)) => (s, alpha),
-            None => (self.screen, self.modal_fade.open_alpha(MODAL_FADE)),
-        };
-        if !matches!(screen, Screen::Home) {
-            cmds.push(DrawCmd::Fill {
-                rect: Rect::new(0, 0, screen_w, screen_h),
-                color: crate::ui::render::Color::RGBA(0, 0, 0, (f32::from(ui::style::theme().scrim.a) * m) as u8),
-            });
-            let dy = ((1.0 - m) * 26.0) as i32;
-            // The tile now covers only the card region (see `prepare_modal`), so it
-            // composites there rather than full-screen. `pop_in_rect` scaling around this
-            // rect's center is the card's own center — the same visual pop as before.
-            let modal_base = self.modal_tile_region.offset(0, dy);
-            let modal_dst = if closing_frame.is_some() {
-                modal_base
-            } else {
-                ui::animation::pop_in_rect(modal_base, m, MODAL_POP_SHRINK)
-            };
-            cmds.push(DrawCmd::Tex {
-                tile: tile::MODAL,
-                dst: modal_dst,
-                alpha: (255.0 * m) as u8,
-            });
-            // Scrollable content geometry (Settings rows or About document), computed
-            // once and reused. Scrolling crops the full baked tile, never re-rasterizes.
-            let scroll_geom = self.scroll_geometry_for(screen, screen_w, screen_h, fonts);
-            if let Some((total, _, _, content)) = scroll_geom {
-                // About uses a bounded window; for other screens, window_start is 0.
-                let window_start = match screen {
-                    Screen::About => self.content_window.start,
-                    _ => 0,
-                };
-                let stride = self.scroll_stride_for(screen, fonts);
-                // The animated offset (see `sync_modal_scroll`), in absolute content pixels,
-                // rebased onto whatever slice is currently baked into the tile.
-                let scroll_px = self
-                    .modal_scroll_px
-                    .clamp(0, Self::max_scroll_px(total, stride, content.height()));
-                let src_y = scroll_px - window_start as i32 * stride;
-                cmds.push(DrawCmd::TexCropped {
-                    tile: tile::SCROLL_CONTENT,
-                    src: Rect::new(0, src_y, content.width(), content.height()),
-                    dst: content.offset(0, dy),
-                    alpha: (255.0 * m) as u8,
-                });
-                // Bottom fade, only while rows remain below the viewport — it is the
-                // "there is more" signal, so it has to vanish exactly when scrolling has
-                // reached the end, or it reads as content that can never be got to.
-                //
-                // Pushed here, between the content and the focused-row tile below, on
-                // purpose: focus must never look dimmed just because it sits on the last
-                // visible row, and an open dropdown (pushed next) must cover the band
-                // rather than show through it.
-                // Keyed off pixels, not rows: at either end of the list the offset is clamped
-                // mid-row, so a row-based test would keep claiming there is more beyond.
-                let fade_h = ui::widgets::SCROLL_FADE_H.min(content.height());
-                if scroll_px > 0 {
-                    cmds.push(DrawCmd::Tex {
-                        tile: tile::SCROLL_FADE_TOP,
-                        dst: Rect::new(content.x(), content.y() + dy, content.width(), fade_h),
-                        alpha: (255.0 * m) as u8,
-                    });
-                }
-                if scroll_px < Self::max_scroll_px(total, stride, content.height()) {
-                    cmds.push(DrawCmd::Tex {
-                        tile: tile::SCROLL_FADE,
-                        dst: Rect::new(
-                            content.x(),
-                            content.y() + dy + (content.height() - fade_h) as i32,
-                            content.width(),
-                            fade_h,
-                        ),
-                        alpha: (255.0 * m) as u8,
-                    });
-                }
-            }
-            // Dropdown overlay (Settings or Diagnostics).
-            if let Some((row, _, dd_alpha)) = self.dropdown_draw_state() {
-                if let Some((content, scroll_px)) = self.dropdown_geom(screen, screen_w, screen_h, fonts) {
-                    let overlay_rect = view::settings::dropdown_overlay_rect_at_px(content, row, scroll_px);
-                    let options_len = match screen {
-                        Screen::Diagnostics => menu::LOG_LEVEL_OPTIONS.len(),
-                        _ => menu::dropdown_option_count(menu::settings_logical_row(&self.settings, row)),
-                    };
-                    cmds.push(DrawCmd::Tex {
-                        tile: tile::DROPDOWN_OVERLAY,
-                        dst: Rect::new(
-                            overlay_rect.x(),
-                            overlay_rect.y() + dy,
-                            overlay_rect.width(),
-                            options_len as u32 * ui::widgets::DROPDOWN_OPTION_H,
-                        ),
-                        alpha: (255.0 * m * dd_alpha) as u8,
-                    });
-                }
-            }
-            // Focused widget of the active modal (setting row, button, etc.);
-            // composites on shell at its on-screen position (no re-rasterize on move).
-            //
-            // Skipped entirely once the modal is closing: the position is recomputed
-            // from live per-screen state, which Back may have already torn down (e.g.
-            // `host_menu_index` cleared, collapsing the host-menu card to the screen
-            // centre and floating the highlight there). The shell and scroll-content
-            // tiles still render the focused row through the fade, so dropping just the
-            // zoom-highlight overlay is invisible — and correct.
-            let focus_rect = if closing_frame.is_some() {
-                None
-            } else {
-                match screen {
-                    Screen::Settings => {
-                        let (total, _, _, content) = scroll_geom.expect("screen is Screen::Settings");
-                        // Positioned from the animated pixel offset, not the row index: the baked
-                        // list is cropped at that offset, and the focus tile *is* the focused row
-                        // re-rendered — so anchoring it to the quantized row would show that row's
-                        // content twice, in two places, for the length of every scroll.
-                        let stride = ui::widgets::focus_row_stride() as i32;
-                        let px = self
-                            .modal_scroll_px
-                            .clamp(0, Self::max_scroll_px(total, stride, content.height()));
-                        Some(ui::widgets::focus_row_rect_at_px(content, self.settings_focused, px))
-                    }
-                    Screen::Wake => self.wake.as_ref().filter(|w| !w.mac.is_empty()).map(|w| {
-                        Self::confirm_focus_button_rect(
-                            screen_w,
-                            screen_h,
-                            fonts,
-                            &view::wake::status_text(w),
-                            w.focused,
-                        )
-                    }),
-                    Screen::Pairing => {
-                        let card = view::pairing::card_rect(screen_w, screen_h, fonts);
-                        Some(match self.pairing_focus {
-                            PairingFocus::Pin => {
-                                let digit_y = view::pairing::pin_row_y(card, fonts);
-                                view::pairing::digit_rect(card, digit_y, self.pin_digit_index)
-                            }
-                            PairingFocus::RequestAccess => view::pairing::request_button_rect(card, fonts),
-                        })
-                    }
-                    Screen::ForgetHost => {
-                        let name = self
-                            .host_menu_index
-                            .and_then(|i| self.entries.get(i))
-                            .map(HostEntry::name)
-                            .unwrap_or_default();
-                        Some(Self::confirm_focus_button_rect(
-                            screen_w,
-                            screen_h,
-                            fonts,
-                            &view::forget::subtitle(name),
-                            self.host_menu_focused,
-                        ))
-                    }
-                    Screen::HostMenu => {
-                        let subtitle = self.host_menu_subtitle();
-                        let rows = self.host_menu_actions().len();
-                        let card = view::hostmenu::card_rect(screen_w, screen_h, fonts, &subtitle, rows);
-                        let content = ui::widgets::list_modal_content_rect(card, fonts, &subtitle, rows);
-                        Some(ui::widgets::focus_row_rect(content, self.menu_focused))
-                    }
-                    Screen::WakeSettings => {
-                        let card = view::wakesettings::card_rect(screen_w, screen_h, fonts);
-                        let content = ui::widgets::list_modal_content_rect(
-                            card,
-                            fonts,
-                            view::wakesettings::SUBTITLE,
-                            menu::DIAGNOSTICS_ROW_COUNT,
-                        );
-                        Some(ui::widgets::focus_row_rect(content, self.wake_settings_focused))
-                    }
-                    Screen::SpeedTest => view::speedtest::finished(self.speed_test.as_ref()).then(|| {
-                        let card = view::speedtest::card_rect(
-                            screen_w,
-                            screen_h,
-                            fonts,
-                            self.speed_test.as_ref(),
-                            &self.speed_test_name,
-                        );
-                        ui::widgets::confirm_button_rect(
-                            view::speedtest::buttons_rect(card, fonts, self.speed_test.as_ref(), &self.speed_test_name),
-                            self.speed_test_focused,
-                        )
-                    }),
-                    Screen::Diagnostics => {
-                        let card = view::diagnostics::card_rect(screen_w, screen_h, fonts);
-                        let content = ui::widgets::list_modal_content_rect(
-                            card,
-                            fonts,
-                            view::diagnostics::SUBTITLE,
-                            menu::DIAGNOSTICS_ROW_COUNT,
-                        );
-                        Some(ui::widgets::focus_row_rect(content, self.diagnostics_focused))
-                    }
-                    Screen::Experimental => {
-                        let rows = view::experimental::row_count(Self::rooted());
-                        let card = view::experimental::card_rect(screen_w, screen_h, fonts, Self::rooted());
-                        let content =
-                            ui::widgets::list_modal_content_rect(card, fonts, view::experimental::SUBTITLE, rows);
-                        Some(ui::widgets::focus_row_rect(content, self.experimental_focused))
-                    }
-                    Screen::CursorSettings => {
-                        let card = view::cursorsettings::card_rect(screen_w, screen_h, fonts);
-                        let content = ui::widgets::list_modal_content_rect(
-                            card,
-                            fonts,
-                            view::cursorsettings::SUBTITLE,
-                            menu::CURSOR_ROW_COUNT,
-                        );
-                        Some(ui::widgets::focus_row_rect(content, self.cursor_settings_focused))
-                    }
-                    Screen::SendLogs => Some(Self::confirm_focus_button_rect(
-                        screen_w,
-                        screen_h,
-                        fonts,
-                        view::sendlogs::SUBTITLE,
-                        self.send_logs_focused,
-                    )),
-                    Screen::Home | Screen::AddHost | Screen::EditHost | Screen::About | Screen::PinLimit => None,
-                }
-            };
-            if let Some(rect) = focus_rect {
-                let pad = ui::tiles::ROW_TILE_PAD;
-                let base = Rect::new(
-                    rect.x() - pad,
-                    rect.y() - pad + dy,
-                    rect.width() + 2 * pad as u32,
-                    rect.height() + 2 * pad as u32,
-                );
-                // The zoom-in: same GPU-scale-around-center technique as the
-                // grid's card focus pop (see above) — `modal_focus_tile` is
-                // rasterized once at its literal size, never re-rendered for
-                // this (except while `switch_anim` animates its content, see
-                // `prepare_tiles`).
-                let f = ui::animation::anim_frac(self.modal_focus_anim, ui::animation::FOCUS_POP);
-                let dst = ui::animation::zoom_rect(base, f, 0.02);
-                let alpha = (255.0 * m) as u8;
-                // In a scrolling modal the focused row can hang past the viewport's bottom
-                // edge mid-glide (the crop lags the row offset by up to one stride), so it is
-                // clipped rather than left to paint over the card's chrome. Every other modal
-                // keeps the plain unclipped path — none of them scrolls.
-                let tile_size = tiles.get(tile::MODAL_FOCUS).map(|p| (p.width(), p.height()));
-                match (scroll_geom, tile_size) {
-                    (Some((_, _, _, content)), Some((tw, th))) => {
-                        let viewport = Rect::new(
-                            content.x() - pad,
-                            content.y() - pad + dy,
-                            content.width() + 2 * pad as u32,
-                            content.height() + 2 * pad as u32,
-                        );
-                        if let Some((src, visible)) = Self::clip_tile(dst, viewport, tw, th) {
-                            cmds.push(DrawCmd::TexCropped {
-                                tile: tile::MODAL_FOCUS,
-                                src,
-                                dst: visible,
-                                alpha,
-                            });
-                        }
-                    }
-                    _ => cmds.push(DrawCmd::Tex {
-                        tile: tile::MODAL_FOCUS,
-                        dst,
-                        alpha,
-                    }),
-                }
-            }
-            // The open dropdown's focused option — same idea, composited on
-            // top of the shell's unfocused option list at its actual
-            // position, so navigating dropdown options needs no modal
-            // re-rasterize either. `Settings` or `Diagnostics`.
-            if let Some((row, focused, dd_alpha)) = self.dropdown_draw_state() {
-                if let Some((content, scroll_px)) = self.dropdown_geom(screen, screen_w, screen_h, fonts) {
-                    let overlay_rect = view::settings::dropdown_overlay_rect_at_px(content, row, scroll_px);
-                    let option_rect = ui::widgets::dropdown_option_rect(overlay_rect, focused);
-                    cmds.push(DrawCmd::Tex {
-                        tile: tile::DROPDOWN_FOCUS,
-                        dst: Rect::new(
-                            option_rect.x(),
-                            option_rect.y() + dy,
-                            option_rect.width(),
-                            option_rect.height(),
-                        ),
-                        alpha: (255.0 * m * dd_alpha) as u8,
-                    });
-                }
-            }
-            // Whichever modal is scrollable, its indicator — full opacity for
-            // `SCROLL_INDICATOR_HOLD`, then a linear fade over `SCROLL_INDICATOR_FADE`
-            // (names kept from when only Settings had one; every scrollable modal now
-            // shares the same timing and the same `self.scroll.shown_at` clock, since
-            // only one is ever open at a time).
-            if let Some((total, visible, card, content)) = scroll_geom {
-                if total > visible {
-                    let scroll_alpha = self.scroll.shown_at.map_or(0.0, |t| {
-                        let elapsed = t.elapsed();
-                        if elapsed < SCROLL_INDICATOR_HOLD {
-                            1.0
-                        } else {
-                            let fading = (elapsed - SCROLL_INDICATOR_HOLD).as_secs_f32();
-                            1.0 - (fading / SCROLL_INDICATOR_FADE.as_secs_f32()).clamp(0.0, 1.0)
-                        }
-                    });
-                    if scroll_alpha > 0.0 {
-                        // Sits nearer the card's edge than the content's, so it doesn't
-                        // overlap a Settings row's dropdown pill/slider/switch. The `26`
-                        // offset isn't derived from either modal's own width fraction —
-                        // re-check both if either changes.
-                        let dst = Rect::new(
-                            card.right() - 26,
-                            content.y() + dy,
-                            SCROLL_INDICATOR_TILE_W,
-                            content.height(),
-                        );
-                        cmds.push(DrawCmd::Tex {
-                            tile: tile::SCROLL_INDICATOR,
-                            dst,
-                            alpha: (255.0 * m * scroll_alpha) as u8,
-                        });
-                    }
-                }
-            }
-        }
-    }
-
-    /// Sidebar family compose: the focused-row highlight overlay (the strip itself
-    /// is an unconditional `tile::SIDEBAR` blit in `draw_list`). Reads only the
-    /// `RenderInput` slice — a template for the per-family `TileCache::compose` split.
-    fn compose_sidebar_focus(input: &render_input::RenderInput<'_>, screen_h: u32, cmds: &mut Vec<DrawCmd>) {
-        let sidebar_focus_row = match input.home_focus {
-            HomeFocus::Sidebar(i) | HomeFocus::SidebarMenu(i) => Some(i),
-            HomeFocus::Grid(_) => None,
-        };
-        if let Some(i) = sidebar_focus_row {
-            let rect = view::sidebar::nav_row_rect(i, input.entries.len() + 2, screen_h);
-            let pad = ui::tiles::ROW_TILE_PAD;
-            cmds.push(DrawCmd::Tex {
-                tile: tile::FOCUS_ROW,
-                dst: Rect::new(
-                    rect.x() - pad,
-                    rect.y() - pad,
-                    rect.width() + 2 * pad as u32,
-                    rect.height() + 2 * pad as u32,
-                ),
-                alpha: 0xff,
-            });
-        }
-    }
-
-    /// Grid family compose: the card tiles at their scrolled positions, the pinned
-    /// divider, and the focused card with its ring/outline/pin-badge pop. Only reached
-    /// once the grid is revealed. Extracted from `draw_list` (A2 staging).
-    #[allow(clippy::too_many_arguments)]
-    fn compose_grid(&self, screen_h: u32, grid_x: i32, available_w: u32, columns: usize, cmds: &mut Vec<DrawCmd>) {
-        let count = self.grid_len(columns);
-        let focused = match self.home_focus {
-            HomeFocus::Grid(i) if i < count => Some(i),
-            HomeFocus::Grid(_) | HomeFocus::Sidebar(_) | HomeFocus::SidebarMenu(_) => None,
-        };
-        let pad = ui::tiles::CARD_TILE_PAD;
-        let layout = self.grid_layout(columns);
-        for idx in 0..count {
-            if Some(idx) == focused {
-                continue; // drawn last, on top of its neighbors
-            }
-            // padding after a partial pinned row — nothing to draw
-            let Some(pin_id) = layout.pin_id_at(&self.games, idx) else {
-                continue;
-            };
-            let r = self.scrolled_card_rect(idx, columns, grid_x, available_w);
-            if r.bottom() + pad < 0 || r.y() - pad > screen_h as i32 {
-                continue; // culled — fully off-screen at this scroll offset
-            }
-            let Some(card) = self.card_ids.get(pin_id) else {
-                continue; // not rasterized yet — outside the build window
-            };
-            // A card that just landed is still zooming up to full size.
-            let pop = self.card_pop_frac(pin_id);
-            let base = Rect::new(
-                r.x() - pad,
-                r.y() - pad,
-                r.width() + 2 * pad as u32,
-                r.height() + 2 * pad as u32,
-            );
-            cmds.push(DrawCmd::Tex {
-                tile: card,
-                dst: ui::animation::pop_in_rect(base, pop, CARD_POP_SHRINK),
-                alpha: (255.0 * pop) as u8,
-            });
-        }
-        // The divider between pinned games and the rest — scrolled with
-        // everything else (there's no separate fixed region), so it's just
-        // another rect at its own scrolled position, culled the same way.
-        if let Some(sep) = self.pinned_separator_rect(columns, grid_x, available_w) {
-            if sep.y() >= 0 && sep.y() <= screen_h as i32 {
-                cmds.push(DrawCmd::Fill {
-                    rect: sep,
-                    color: crate::ui::render::Color::RGBA(0xff, 0xff, 0xff, 0x20),
-                });
-            }
-        }
-        if let Some(idx) = focused {
-            if let Some(pin_id) = layout.pin_id_at(&self.games, idx) {
-                // The focus pop: the GPU scales the (unfocused) card tile up
-                // around its center as the pop progresses, with the shared glow
-                // tile fading in behind it at the same scale.
-                let f = ui::animation::anim_frac(self.focus_anim, ui::animation::FOCUS_POP);
-                let r = self.scrolled_card_rect(idx, columns, grid_x, available_w);
-                let card_base = Rect::new(
-                    r.x() - pad,
-                    r.y() - pad,
-                    r.width() + 2 * pad as u32,
-                    r.height() + 2 * pad as u32,
-                );
-                let Some(card) = self.card_ids.get(pin_id) else {
-                    return; // not rasterized yet
-                };
-                let pop = self.card_pop_frac(pin_id);
-                let popped = |base: Rect| ui::animation::pop_in_rect(base, pop, CARD_POP_SHRINK);
-                // Glow drawn first — it's a halo behind the card, not an outline
-                // on top of it.
-                let rp = ui::tiles::FOCUS_RING_PAD;
-                let ring_base = Rect::new(
-                    r.x() - rp,
-                    r.y() - rp,
-                    r.width() + 2 * rp as u32,
-                    r.height() + 2 * rp as u32,
-                );
-                cmds.push(DrawCmd::Tex {
-                    tile: tile::RING,
-                    dst: popped(ui::animation::zoom_rect(ring_base, f, CARD_GROWTH)),
-                    alpha: (255.0 * f * pop) as u8,
-                });
-                // The focused card zooms in on first appearance like any other,
-                // composed with its focus pop — both scale around the card's own
-                // center, so they can't fight over position.
-                cmds.push(DrawCmd::Tex {
-                    tile: card,
-                    dst: popped(ui::animation::zoom_rect(card_base, f, CARD_GROWTH)),
-                    alpha: (255.0 * pop) as u8,
-                });
-                // The crisp outline, on top of the card art — a clean edge
-                // between it and the glow behind, unlike the glow's own
-                // soft, blurred boundary.
-                let op = ui::tiles::CARD_OUTLINE_PAD;
-                let outline_base = Rect::new(
-                    r.x() - op,
-                    r.y() - op,
-                    r.width() + 2 * op as u32,
-                    r.height() + 2 * op as u32,
-                );
-                cmds.push(DrawCmd::Tex {
-                    tile: tile::CARD_OUTLINE,
-                    dst: popped(ui::animation::zoom_rect(outline_base, f, CARD_GROWTH)),
-                    alpha: (255.0 * f * pop) as u8,
-                });
-                if self.selected_known_host().is_some_and(|h| h.is_pinned(pin_id)) {
-                    let badge = ui::tiles::PIN_BADGE_SIZE;
-                    let badge_base = Rect::new(
-                        r.right() - badge as i32 - PIN_BADGE_MARGIN,
-                        r.y() + PIN_BADGE_MARGIN,
-                        badge,
-                        badge,
-                    );
-                    // Corner-anchored, so it only fades — scaling it around its
-                    // own center would drift it off the shrunken card.
-                    cmds.push(DrawCmd::Tex {
-                        tile: tile::PIN_BADGE,
-                        dst: ui::animation::zoom_rect(badge_base, f, CARD_GROWTH),
-                        alpha: (255.0 * pop) as u8,
-                    });
-                }
-            }
-        }
-    }
-
-    pub fn draw_list(
-        &self,
-        tiles: &TileStore,
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::text::Fonts,
-    ) -> ui::render::DrawList {
-        let input = self.render_input();
-        let mut cmds = Vec::new();
-        let grid_x = ui::widgets::SIDEBAR_W as i32;
-        let available_w = screen_w.saturating_sub(ui::widgets::SIDEBAR_W);
-        let columns = view::home::grid_columns(available_w);
-
-        cmds.push(DrawCmd::Tex {
-            tile: tile::SIDEBAR,
-            dst: Rect::new(0, 0, ui::widgets::SIDEBAR_W, screen_h),
-            alpha: 0xff,
-        });
-
-        if !input.host_selected {
-            if let Some(p) = tiles.get(tile::NO_HOST) {
-                cmds.push(DrawCmd::Tex {
-                    tile: tile::NO_HOST,
-                    dst: Rect::new(
-                        grid_x + view::home::GRID_PAD,
-                        view::home::GRID_TOP_Y,
-                        p.width(),
-                        p.height(),
-                    ),
-                    alpha: 0xff,
-                });
-            }
-        } else if !input.grid_reveal_ready {
-            let phase = self.spinner_since.map_or(0.0, |s| s.elapsed().as_secs_f32());
-            let (idx, frame) = crate::assets::spinner_frame_at(phase);
-            let x = grid_x + (available_w as i32 - frame.width as i32) / 2;
-            // 40% down rather than dead-center, which reads as slightly low on a TV.
-            let area_h = screen_h as i32 - view::home::GRID_TOP_Y;
-            let y = view::home::GRID_TOP_Y + (area_h - frame.height as i32) * 2 / 5;
-            cmds.push(DrawCmd::Tex {
-                tile: tile::spinner(idx),
-                dst: Rect::new(x, y, frame.width, frame.height),
-                alpha: 0xff,
-            });
-        } else {
-            self.compose_grid(screen_h, grid_x, available_w, columns, &mut cmds);
-        }
-        if input.has_status {
-            if let Some(p) = tiles.get(tile::STATUS) {
-                let line_h = fonts.raster.height(fonts.label) + 6;
-                let box_h = 2 * line_h as u32 + 2 * STATUS_BG_PAD as u32;
-                let box_y = screen_h as i32 - box_h as i32;
-                cmds.push(DrawCmd::Fill {
-                    rect: Rect::new(grid_x, box_y, available_w, box_h),
-                    color: ui::style::theme().scrim,
-                });
-                let y = box_y + (box_h as i32 - p.height() as i32) / 2;
-                cmds.push(DrawCmd::Tex {
-                    tile: tile::STATUS,
-                    dst: Rect::new(grid_x + view::home::GRID_PAD, y, p.width(), p.height()),
-                    alpha: 0xff,
-                });
-            }
-        }
-
-        Self::compose_sidebar_focus(&input, screen_h, &mut cmds);
-
-        self.compose_modal(tiles, screen_w, screen_h, fonts, &mut cmds);
-        // The launch transition: the confirmed card zooms in around its own
-        // center (same `zoom_rect` technique as the focus pop, so its aspect
-        // ratio never changes) while a black scrim blends in over it, both driven
-        // by the same clock — the card keeps zooming for the whole fade.
-        if let (Some(t), Some(idx)) = (self.launch_anim, self.launch_anim_idx) {
-            let f = ui::animation::anim_frac(Some(t), hero::LAUNCH_FADE);
-            let base = self.scrolled_card_rect(idx, columns, grid_x, available_w);
-            if let Some(card) = self
-                .pin_id_at_grid_idx(idx, columns)
-                .and_then(|pin_id| self.card_ids.get(pin_id))
-            {
-                cmds.push(DrawCmd::Tex {
-                    tile: card,
-                    dst: ui::animation::zoom_rect(base, f, LAUNCH_GROWTH),
-                    alpha: 0xff,
-                });
-            }
-            cmds.push(DrawCmd::Fill {
-                rect: Rect::new(0, 0, screen_w, screen_h),
-                color: crate::ui::render::Color::RGBA(0, 0, 0, (255.0 * f) as u8),
-            });
-            // With wide art for this game, the loading screen is that art instead of the
-            // bare black: it fades in over the scrim above (so a hero arriving mid-
-            // handshake still eases in rather than snapping), then drifts slowly left to
-            // right for as long as the stream takes to come up.
-            self.compose_hero(screen_w, screen_h, &mut cmds);
-        }
-        cmds
-    }
-
-    /// The connecting screen's hero backdrop: fade in/out, slow pan, and its dimming
-    /// scrim. Fading rather than cutting at both ends, over the same black the launch faded
-    /// to, so a hero arriving mid-handshake eases in and live video arrives out of black
-    /// rather than from a lit image.
-    fn compose_hero(&self, screen_w: u32, screen_h: u32, cmds: &mut ui::render::DrawList) {
-        let Some(hero) = self.hero.visible() else { return };
-        let f = self.hero.opacity();
-        cmds.push(DrawCmd::TexF {
-            tile: tile::HERO,
-            dst: hero::hero_pan_dst(hero.width, hero.height, screen_w, screen_h, self.hero.panned_for()),
-            alpha: (255.0 * f) as u8,
-        });
-        cmds.push(DrawCmd::Fill {
-            rect: Rect::new(0, 0, screen_w, screen_h),
-            color: crate::ui::render::Color::RGBA(0, 0, 0, (hero::HERO_SCRIM_ALPHA * f) as u8),
-        });
     }
 }
